@@ -2,10 +2,12 @@ repositories {
     maven {url "https://hub.spigotmc.org/nexus/content/groups/public/"}
     maven { url = "https://oss.sonatype.org/content/repositories/snapshots/"}
     maven {url "http://nexus.hc.to/content/repositories/pub_releases"}
+    maven { url = "https://repo.codemc.org/repository/maven-public" }
     mavenLocal()
 }
 
 dependencies {
+    implementation project (':Core')
     compile project(':Core')
     compile 'org.spigotmc:spigot-api:1.12.1-R0.1-SNAPSHOT'
     compile("net.milkbowl.vault:VaultAPI:1.6") {
@@ -13,8 +15,8 @@ dependencies {
     }
 }
 
-sourceCompatibility = 1.7
-targetCompatibility = 1.7
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
 
 processResources {
     from('src/main/resources') {
@@ -32,7 +34,6 @@ shadowJar {
     dependencies {
         include(dependency(':Core'))
     }
-    relocate('org.mcstats', 'com.plotsquared.stats')
     archiveName = "${parent.name}-${project.name}-${parent.version}.jar"
     destinationDir = file '../target'
 }
