@@ -4,8 +4,8 @@ dependencies {
     compile 'com.google.code.gson:gson:2.2.4'
 }
 
-sourceCompatibility = 1.7
-targetCompatibility = 1.7
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
 
 
 processResources {
@@ -20,7 +20,7 @@ processResources {
 
 jar.archiveName="plotsquared-api-${project.parent.version}.jar"
 jar.destinationDir = file '../mvn/com/plotsquared/plotsquared-api/' + project.parent.version
-task createPom << {
+task createPom {
     pom {
         project {
             groupId 'com.plotsquared'
