@@ -67,12 +67,7 @@ public final class BukkitMain extends JavaPlugin implements Listener, IPlotMain 
                 pluginsField.setAccessible(true);
                 lookupNamesField.setAccessible(true);
                 List<Plugin> plugins = (List<Plugin>) pluginsField.get(manager);
-                Iterator<Plugin> iter = plugins.iterator();
-                while (iter.hasNext()) {
-                    if (iter.next().getName().startsWith("PlotMe")) {
-                        iter.remove();
-                    }
-                }
+                plugins.removeIf(plugin -> plugin.getName().startsWith("PlotMe"));
                 Map<String, Plugin> lookupNames = (Map<String, Plugin>) lookupNamesField.get(manager);
                 lookupNames.remove("PlotMe");
                 lookupNames.remove("PlotMe-DefaultGenerator");
@@ -106,6 +101,7 @@ public final class BukkitMain extends JavaPlugin implements Listener, IPlotMain 
     private int[] version;
     private String name;
     private SingleWorldListener singleWorldListener;
+    private boolean metricsStarted;
 
     @Override
     public int[] getServerVersion() {
@@ -135,25 +131,24 @@ public final class BukkitMain extends JavaPlugin implements Listener, IPlotMain 
 
     @Override
     public void onEnable() {
+        if (getServerVersion()[1] > 12) {
+            Bukkit.getLogger().severe("================================================");
+            Bukkit.getLogger().severe("====== THIS PLOTSQUARED VERSION IS NOT ======");
+            Bukkit.getLogger().severe("======  FOR USE ON MINECRAFT 1.13 OR   ======");
+            Bukkit.getLogger().severe("======     ABOVE. DISABLING PLUGIN.    ======");
+            Bukkit.getLogger().severe("======          DOWNLOAD FROM          ======");
+            Bukkit.getLogger().severe("https://ci.athion.net/job/PlotSquared-Breaking/");
+            Bukkit.getLogger().severe("==============================================");
+            Bukkit.getPluginManager().disablePlugin(this);
+            throw new UnsupportedOperationException("INCORRECT PLOTSQUARED VERSION");
+        }
         if (pluginMap != null) {
             pluginMap.put("PlotMe-DefaultGenerator", this);
         }
         this.name = getDescription().getName();
-        getServer().getName();
         new PS(this, "Bukkit");
-        if (Settings.Enabled_Components.METRICS) {
-            new Metrics(this).start();
-            PS.log(C.PREFIX + "&6Metrics enabled.");
-        } else {
-            PS.log(C.CONSOLE_PLEASE_ENABLE_METRICS.f(getPluginName()));
-        }
         if (Settings.Enabled_Components.WORLDS) {
-            TaskManager.IMP.taskRepeat(new Runnable() {
-                @Override
-                public void run() {
-                    unload();
-                }
-            }, 20);
+            TaskManager.IMP.taskRepeat(this::unload, 20);
             try {
                 singleWorldListener = new SingleWorldListener(this);
             } catch (Exception e) {
@@ -176,8 +171,8 @@ public final class BukkitMain extends JavaPlugin implements Listener, IPlotMain 
                 ReflectionUtils.RefClass classCraftWorld = getRefClass("{cb}.CraftWorld");
                 methodUnloadChunk0 = classCraftWorld.getRealClass().getDeclaredMethod("unloadChunk0", int.class, int.class, boolean.class);
                 methodUnloadChunk0.setAccessible(true);
-            } catch (Throwable ignore) {
-                ignore.printStackTrace();
+            } catch (Throwable throwable) {
+                throwable.printStackTrace();
             }
         }
         PlotAreaManager manager = PS.get().getPlotAreaManager();
@@ -227,8 +222,10 @@ public final class BukkitMain extends JavaPlugin implements Listener, IPlotMain 
                                     if (!result) {
                                         continue outer;
                                     }
-                                } while (index < chunks.length && System.currentTimeMillis() - start < 5);
-                                return;
+                                    if (System.currentTimeMillis() - start > 5) {
+                                        return;
+                                    }
+                                } while (index < chunks.length);
                             }
                         }
                     }
@@ -307,218 +304,213 @@ public final class BukkitMain extends JavaPlugin implements Listener, IPlotMain 
     @Override
     public void runEntityTask() {
         PS.log(C.PREFIX + "KillAllEntities started.");
-        TaskManager.runTaskRepeat(new Runnable() {
+        TaskManager.runTaskRepeat(() -> PS.get().foreachPlotArea(new RunnableVal<PlotArea>() {
             @Override
-            public void run() {
-                PS.get().foreachPlotArea(new RunnableVal<PlotArea>() {
-                    @Override
-                    public void run(PlotArea plotArea) {
-                        World world = Bukkit.getWorld(plotArea.worldname);
-                        try {
-                            if (world == null) {
-                                return;
+            public void run(PlotArea plotArea) {
+                World world = Bukkit.getWorld(plotArea.worldname);
+                try {
+                    if (world == null) {
+                        return;
+                    }
+                    List<Entity> entities = world.getEntities();
+                    Iterator<Entity> iterator = entities.iterator();
+                    while (iterator.hasNext()) {
+                        Entity entity = iterator.next();
+                        switch (entity.getType()) {
+                            case EGG:
+                            case COMPLEX_PART:
+                            case FISHING_HOOK:
+                            case ENDER_SIGNAL:
+                            case LINGERING_POTION:
+                            case AREA_EFFECT_CLOUD:
+                            case EXPERIENCE_ORB:
+                            case LEASH_HITCH:
+                            case FIREWORK:
+                            case WEATHER:
+                            case LIGHTNING:
+                            case WITHER_SKULL:
+                            case UNKNOWN:
+                            case PLAYER:
+                                // non moving / unmovable
+                                continue;
+                            case THROWN_EXP_BOTTLE:
+                            case SPLASH_POTION:
+                            case SNOWBALL:
+                            case SHULKER_BULLET:
+                            case SPECTRAL_ARROW:
+                            case TIPPED_ARROW:
+                            case ENDER_PEARL:
+                            case ARROW:
+                            case LLAMA_SPIT:
+                                // managed elsewhere | projectile
+                                continue;
+                            case ITEM_FRAME:
+                            case PAINTING:
+                                // Not vehicles
+                                continue;
+                            case ARMOR_STAND:
+                                // Temporarily classify as vehicle
+                            case MINECART:
+                            case MINECART_CHEST:
+                            case MINECART_COMMAND:
+                            case MINECART_FURNACE:
+                            case MINECART_HOPPER:
+                            case MINECART_MOB_SPAWNER:
+                            case ENDER_CRYSTAL:
+                            case MINECART_TNT:
+                            case BOAT:
+                                if (Settings.Enabled_Components.KILL_ROAD_VEHICLES) {
+                                    com.intellectualcrafters.plot.object.Location location = BukkitUtil.getLocation(entity.getLocation());
+                                    Plot plot = location.getPlot();
+                                    if (plot == null) {
+                                        if (location.isPlotArea()) {
+                                            iterator.remove();
+                                            entity.remove();
+                                        }
+                                        continue;
+                                    }
+                                    List<MetadataValue> meta = entity.getMetadata("plot");
+                                    if (meta.isEmpty()) {
+                                        continue;
+                                    }
+                                    Plot origin = (Plot) meta.get(0).value();
+                                    if (!plot.equals(origin.getBasePlot(false))) {
+                                        iterator.remove();
+                                        entity.remove();
+                                    }
+                                    continue;
+                                } else {
+                                    continue;
+                                }
+                            case SMALL_FIREBALL:
+                            case FIREBALL:
+                            case DRAGON_FIREBALL:
+                            case DROPPED_ITEM:
+                                if (Settings.Enabled_Components.KILL_ROAD_ITEMS && plotArea.getOwnedPlotAbs(BukkitUtil.getLocation(entity.getLocation())) == null) {
+                                    entity.remove();
+                                }
+                                // dropped item
+                                continue;
+                            case PRIMED_TNT:
+                            case FALLING_BLOCK:
+                                // managed elsewhere
+                                continue;
+                            case LLAMA:
+                            case DONKEY:
+                            case MULE:
+                            case ZOMBIE_HORSE:
+                            case SKELETON_HORSE:
+                            case HUSK:
+                            case ELDER_GUARDIAN:
+                            case WITHER_SKELETON:
+                            case STRAY:
+                            case ZOMBIE_VILLAGER:
+                            case EVOKER:
+                            case EVOKER_FANGS:
+                            case VEX:
+                            case VINDICATOR:
+                            case POLAR_BEAR:
+                            case BAT:
+                            case BLAZE:
+                            case CAVE_SPIDER:
+                            case CHICKEN:
+                            case COW:
+                            case CREEPER:
+                            case ENDERMAN:
+                            case ENDERMITE:
+                            case ENDER_DRAGON:
+                            case GHAST:
+                            case GIANT:
+                            case GUARDIAN:
+                            case HORSE:
+                            case IRON_GOLEM:
+                            case MAGMA_CUBE:
+                            case MUSHROOM_COW:
+                            case OCELOT:
+                            case PIG:
+                            case PIG_ZOMBIE:
+                            case RABBIT:
+                            case SHEEP:
+                            case SILVERFISH:
+                            case SKELETON:
+                            case SLIME:
+                            case SNOWMAN:
+                            case SPIDER:
+                            case SQUID:
+                            case VILLAGER:
+                            case WITCH:
+                            case WITHER:
+                            case WOLF:
+                            case ZOMBIE:
+                            default: {
+                                if (Settings.Enabled_Components.KILL_ROAD_MOBS) {
+                                    Location location = entity.getLocation();
+                                    if (BukkitUtil.getLocation(location).isPlotRoad()) {
+                                        if (entity instanceof LivingEntity) {
+                                            LivingEntity livingEntity = (LivingEntity) entity;
+                                            if (!livingEntity.isLeashed() || !entity.hasMetadata("keep")) {
+                                                Entity passenger = entity.getPassenger();
+                                                if (!(passenger instanceof Player) && entity.getMetadata("keep").isEmpty()) {
+                                                    iterator.remove();
+                                                    entity.remove();
+                                                    continue;
+                                                }
+                                            }
+                                        } else {
+                                            Entity passenger = entity.getPassenger();
+                                            if (!(passenger instanceof Player) && entity.getMetadata("keep").isEmpty()) {
+                                                iterator.remove();
+                                                entity.remove();
+                                                continue;
+                                            }
+                                        }
+                                    }
+                                }
+                                continue;
                             }
-                            List<Entity> entities = world.getEntities();
-                            Iterator<Entity> iterator = entities.iterator();
-                            while (iterator.hasNext()) {
-                                Entity entity = iterator.next();
-                                switch (entity.getType()) {
-                                    case EGG:
-                                    case COMPLEX_PART:
-                                    case FISHING_HOOK:
-                                    case ENDER_SIGNAL:
-                                    case LINGERING_POTION:
-                                    case AREA_EFFECT_CLOUD:
-                                    case EXPERIENCE_ORB:
-                                    case LEASH_HITCH:
-                                    case FIREWORK:
-                                    case WEATHER:
-                                    case LIGHTNING:
-                                    case WITHER_SKULL:
-                                    case UNKNOWN:
-                                    case PLAYER:
-                                        // non moving / unmovable
-                                        continue;
-                                    case THROWN_EXP_BOTTLE:
-                                    case SPLASH_POTION:
-                                    case SNOWBALL:
-                                    case SHULKER_BULLET:
-                                    case SPECTRAL_ARROW:
-                                    case TIPPED_ARROW:
-                                    case ENDER_PEARL:
-                                    case ARROW:
-                                    case LLAMA_SPIT:
-                                        // managed elsewhere | projectile
-                                        continue;
-                                    case ITEM_FRAME:
-                                    case PAINTING:
-                                        // Not vehicles
-                                        continue;
-                                    case ARMOR_STAND:
-                                        // Temporarily classify as vehicle
-                                    case MINECART:
-                                    case MINECART_CHEST:
-                                    case MINECART_COMMAND:
-                                    case MINECART_FURNACE:
-                                    case MINECART_HOPPER:
-                                    case MINECART_MOB_SPAWNER:
-                                    case ENDER_CRYSTAL:
-                                    case MINECART_TNT:
-                                    case BOAT:
-                                        if (Settings.Enabled_Components.KILL_ROAD_VEHICLES) {
-                                            com.intellectualcrafters.plot.object.Location location = BukkitUtil.getLocation(entity.getLocation());
-                                            Plot plot = location.getPlot();
-                                            if (plot == null) {
-                                                if (location.isPlotArea()) {
+                            case SHULKER: {
+                                if (Settings.Enabled_Components.KILL_ROAD_MOBS) {
+                                    LivingEntity livingEntity = (LivingEntity) entity;
+                                    List<MetadataValue> meta = entity.getMetadata("plot");
+                                    if (meta != null && !meta.isEmpty()) {
+                                        if (livingEntity.isLeashed()) continue;
+
+                                        List<MetadataValue> keep = entity.getMetadata("keep");
+                                        if (keep != null && !keep.isEmpty()) continue;
+
+                                        PlotId originalPlotId = (PlotId) meta.get(0).value();
+                                        if (originalPlotId != null) {
+                                            com.intellectualcrafters.plot.object.Location pLoc = BukkitUtil.getLocation(entity.getLocation());
+                                            PlotArea area = pLoc.getPlotArea();
+                                            if (area != null) {
+                                                PlotId currentPlotId = PlotId.of(area.getPlotAbs(pLoc));
+                                                if (!originalPlotId.equals(currentPlotId) && (currentPlotId == null || !area.getPlot(originalPlotId).equals(area.getPlot(currentPlotId)))) {
                                                     iterator.remove();
                                                     entity.remove();
                                                 }
-                                                continue;
-                                            }
-                                            List<MetadataValue> meta = entity.getMetadata("plot");
-                                            if (meta.isEmpty()) {
-                                                continue;
-                                            }
-                                            Plot origin = (Plot) meta.get(0).value();
-                                            if (!plot.equals(origin.getBasePlot(false))) {
-                                                iterator.remove();
-                                                entity.remove();
-                                            }
-                                            continue;
-                                        } else {
-                                            continue;
-                                        }
-                                    case SMALL_FIREBALL:
-                                    case FIREBALL:
-                                    case DRAGON_FIREBALL:
-                                    case DROPPED_ITEM:
-                                        if (Settings.Enabled_Components.KILL_ROAD_ITEMS && plotArea.getOwnedPlotAbs(BukkitUtil.getLocation(entity.getLocation())) == null) {
-                                            entity.remove();
-                                        }
-                                        // dropped item
-                                        continue;
-                                    case PRIMED_TNT:
-                                    case FALLING_BLOCK:
-                                        // managed elsewhere
-                                        continue;
-                                    case LLAMA:
-                                    case DONKEY:
-                                    case MULE:
-                                    case ZOMBIE_HORSE:
-                                    case SKELETON_HORSE:
-                                    case HUSK:
-                                    case ELDER_GUARDIAN:
-                                    case WITHER_SKELETON:
-                                    case STRAY:
-                                    case ZOMBIE_VILLAGER:
-                                    case EVOKER:
-                                    case EVOKER_FANGS:
-                                    case VEX:
-                                    case VINDICATOR:
-                                    case POLAR_BEAR:
-                                    case BAT:
-                                    case BLAZE:
-                                    case CAVE_SPIDER:
-                                    case CHICKEN:
-                                    case COW:
-                                    case CREEPER:
-                                    case ENDERMAN:
-                                    case ENDERMITE:
-                                    case ENDER_DRAGON:
-                                    case GHAST:
-                                    case GIANT:
-                                    case GUARDIAN:
-                                    case HORSE:
-                                    case IRON_GOLEM:
-                                    case MAGMA_CUBE:
-                                    case MUSHROOM_COW:
-                                    case OCELOT:
-                                    case PIG:
-                                    case PIG_ZOMBIE:
-                                    case RABBIT:
-                                    case SHEEP:
-                                    case SILVERFISH:
-                                    case SKELETON:
-                                    case SLIME:
-                                    case SNOWMAN:
-                                    case SPIDER:
-                                    case SQUID:
-                                    case VILLAGER:
-                                    case WITCH:
-                                    case WITHER:
-                                    case WOLF:
-                                    case ZOMBIE:
-                                    default: {
-                                        if (Settings.Enabled_Components.KILL_ROAD_MOBS) {
-                                            Location location = entity.getLocation();
-                                            if (BukkitUtil.getLocation(location).isPlotRoad()) {
-                                                if (entity instanceof LivingEntity) {
-                                                    LivingEntity livingEntity = (LivingEntity) entity;
-                                                    if (!livingEntity.isLeashed() || !entity.hasMetadata("keep")) {
-                                                        Entity passenger = entity.getPassenger();
-                                                        if (!(passenger instanceof Player) && entity.getMetadata("keep").isEmpty()) {
-                                                            iterator.remove();
-                                                            entity.remove();
-                                                            continue;
-                                                        }
-                                                    }
-                                                } else {
-                                                    Entity passenger = entity.getPassenger();
-                                                    if (!(passenger instanceof Player) && entity.getMetadata("keep").isEmpty()) {
-                                                        iterator.remove();
-                                                        entity.remove();
-                                                        continue;
-                                                    }
-                                                }
                                             }
                                         }
-                                        continue;
-                                    }
-                                    case SHULKER: {
-                                        if (Settings.Enabled_Components.KILL_ROAD_MOBS) {
-                                            LivingEntity livingEntity = (LivingEntity) entity;
-                                            List<MetadataValue> meta = entity.getMetadata("plot");
-                                            if (meta != null && !meta.isEmpty()) {
-                                                if (livingEntity.isLeashed()) continue;
-
-                                                List<MetadataValue> keep = entity.getMetadata("keep");
-                                                if (keep != null && !keep.isEmpty()) continue;
-
-                                                PlotId originalPlotId = (PlotId) meta.get(0).value();
-                                                if (originalPlotId != null) {
-                                                    com.intellectualcrafters.plot.object.Location pLoc = BukkitUtil.getLocation(entity.getLocation());
-                                                    PlotArea area = pLoc.getPlotArea();
-                                                    if (area != null) {
-                                                        PlotId currentPlotId = PlotId.of(area.getPlotAbs(pLoc));
-                                                        if (!originalPlotId.equals(currentPlotId) && (currentPlotId == null || !area.getPlot(originalPlotId).equals(area.getPlot(currentPlotId)))) {
-                                                            iterator.remove();
-                                                            entity.remove();
-                                                        }
-                                                    }
-                                                }
-                                            } else {
-                                                //This is to apply the metadata to already spawned shulkers (see EntitySpawnListener.java)
-                                                com.intellectualcrafters.plot.object.Location pLoc = BukkitUtil.getLocation(entity.getLocation());
-                                                PlotArea area = pLoc.getPlotArea();
-                                                if (area != null) {
-                                                    PlotId currentPlotId = PlotId.of(area.getPlotAbs(pLoc));
-                                                    if (currentPlotId != null) {
-                                                        entity.setMetadata("plot", new FixedMetadataValue((Plugin) PS.get().IMP, currentPlotId));
-                                                    }
-                                                }
+                                    } else {
+                                        //This is to apply the metadata to already spawned shulkers (see EntitySpawnListener.java)
+                                        com.intellectualcrafters.plot.object.Location pLoc = BukkitUtil.getLocation(entity.getLocation());
+                                        PlotArea area = pLoc.getPlotArea();
+                                        if (area != null) {
+                                            PlotId currentPlotId = PlotId.of(area.getPlotAbs(pLoc));
+                                            if (currentPlotId != null) {
+                                                entity.setMetadata("plot", new FixedMetadataValue((Plugin) PS.get().IMP, currentPlotId));
                                             }
                                         }
-                                        continue;
                                     }
                                 }
+                                continue;
                             }
-                        } catch (Throwable e) {
-                            e.printStackTrace();
                         }
                     }
-                });
+                } catch (Throwable e) {
+                    e.printStackTrace();
+                }
             }
-        }, 20);
+        }), 20);
     }
 
     @Override
@@ -774,8 +766,12 @@ public final class BukkitMain extends JavaPlugin implements Listener, IPlotMain 
 
     @Override
     public void startMetrics() {
-        new Metrics(this).start();
+        if (this.metricsStarted) {
+            return;
+        }
+        Metrics metrics = new Metrics(this);// bstats
         PS.log(C.PREFIX + "&6Metrics enabled.");
+        this.metricsStarted = true;
     }
 
     @Override
@@ -799,8 +795,8 @@ public final class BukkitMain extends JavaPlugin implements Listener, IPlotMain 
                 if (!PS.get().hasPlotArea(worldName)) {
                     SetGenCB.setGenerator(BukkitUtil.getWorld(worldName));
                 }
-            } catch (Exception ignored) {
-                PS.log("Failed to reload world: " + world + " | " + ignored.getMessage());
+            } catch (Exception ex) {
+                PS.log("Failed to reload world: " + world + " | " + ex.getMessage());
                 Bukkit.getServer().unloadWorld(world, false);
                 return;
             }
