@@ -61,8 +61,6 @@ import java.io.InputStream;
 import java.net.MalformedURLException;
 import java.net.URISyntaxException;
 import java.net.URL;
-import java.net.URLConnection;
-import java.nio.file.Files;
 import java.sql.SQLException;
 import java.util.ArrayDeque;
 import java.util.ArrayList;
@@ -95,8 +93,6 @@ public class PS{
     private ILogger logger;
     // Current thread
     private final Thread thread;
-    // Platform / Version / Update URL
-    private Updater updater;
     private PlotVersion version;
     // WorldEdit instance
     public WorldEdit worldedit;
@@ -236,29 +232,14 @@ public class PS{
             }
             // Economy
             if (Settings.Enabled_Components.ECONOMY) {
-                TaskManager.runTask(new Runnable() {
-                    @Override
-                    public void run() {
-                        EconHandler.manager = PS.this.IMP.getEconomyHandler();
-                    }
-                });
+                TaskManager.runTask(() -> EconHandler.manager = PS.this.IMP.getEconomyHandler());
             }
 
             // Check for updates
             if (Settings.Enabled_Components.UPDATER) {
-                updater = new Updater();
-                TaskManager.IMP.taskAsync(new Runnable() {
-                    @Override
-                    public void run() {
-                        updater.update(getPlatform(), getVersion());
-                    }
-                });
-                TaskManager.IMP.taskRepeatAsync(new Runnable() {
-                    @Override
-                    public void run() {
-                        updater.update(getPlatform(), getVersion());
-                    }
-                }, 36000);
+                //updater = new Updater();
+                //TaskManager.IMP.taskAsync(() -> updater.update(getPlatform(), getVersion()));
+                //TaskManager.IMP.taskRepeatAsync(() -> updater.update(getPlatform(), getVersion()), 36000);
             }
 
             // World generators:
@@ -272,19 +253,16 @@ public class PS{
                         this.IMP.setGenerator(world);
                     }
                 }
-                TaskManager.runTaskLater(new Runnable() {
-                    @Override
-                    public void run() {
-                        for (String world : section.getKeys(false)) {
-                            if (world.equals("CheckingPlotSquaredGenerator")) {
-                                continue;
-                            }
-                            if (!WorldUtil.IMP.isWorld(world) && !world.equals("*")) {
-                                debug("&c`" + world + "` was not properly loaded - " + IMP.getPluginName() + " will now try to load it properly: ");
-                                debug("&8 - &7Are you trying to delete this world? Remember to remove it from the settings.yml, bukkit.yml and multiverse worlds.yml");
-                                debug("&8 - &7Your world management plugin may be faulty (or non existent)");
-                                PS.this.IMP.setGenerator(world);
-                            }
+                TaskManager.runTaskLater(() -> {
+                    for (String world : section.getKeys(false)) {
+                        if (world.equals("CheckingPlotSquaredGenerator")) {
+                            continue;
+                        }
+                        if (!WorldUtil.IMP.isWorld(world) && !world.equals("*")) {
+                            debug("&c`" + world + "` was not properly loaded - " + IMP.getPluginName() + " will now try to load it properly: ");
+                            debug("&8 - &7Are you trying to delete this world? Remember to remove it from the settings.yml, bukkit.yml and multiverse worlds.yml");
+                            debug("&8 - &7Your world management plugin may be faulty (or non existent)");
+                            PS.this.IMP.setGenerator(world);
                         }
                     }
                 }, 1);
@@ -358,14 +336,6 @@ public class PS{
         return logger;
     }
 
-    /**
-     * The plugin updater
-     * @return
-     */
-    public Updater getUpdater() {
-        return updater;
-    }
-
     public PlotAreaManager getPlotAreaManager() {
         return manager;
     }
@@ -381,29 +351,23 @@ public class PS{
     }
 
     private void startUuidCatching() {
-        TaskManager.runTaskLater(new Runnable() {
-            @Override
-            public void run() {
-                debug("Starting UUID caching");
-                UUIDHandler.startCaching(new Runnable() {
+        TaskManager.runTaskLater(() -> {
+            debug("Starting UUID caching");
+            UUIDHandler.startCaching(() -> {
+                UUIDHandler.add(new StringWrapper("*"), DBFunc.everyone);
+                foreachPlotRaw(new RunnableVal<Plot>() {
                     @Override
-                    public void run() {
-                        UUIDHandler.add(new StringWrapper("*"), DBFunc.everyone);
-                        foreachPlotRaw(new RunnableVal<Plot>() {
-                            @Override
-                            public void run(Plot plot) {
-                                if (plot.hasOwner() && plot.temp != -1) {
-                                    if (UUIDHandler.getName(plot.owner) == null) {
-                                        UUIDHandler.implementation.unknown.add(plot.owner);
-                                    }
-                                }
+                    public void run(Plot plot) {
+                        if (plot.hasOwner() && plot.temp != -1) {
+                            if (UUIDHandler.getName(plot.owner) == null) {
+                                UUIDHandler.implementation.unknown.add(plot.owner);
                             }
-                        });
-                        startExpiryTasks();
-                        startPlotMeConversion();
+                        }
                     }
                 });
-            }
+                startExpiryTasks();
+                startPlotMeConversion();
+            });
         }, 20);
     }
 
@@ -420,19 +384,16 @@ public class PS{
 
     private void startPlotMeConversion() {
         if (Settings.Enabled_Components.PLOTME_CONVERTER || Settings.PlotMe.CACHE_UUDS) {
-            TaskManager.IMP.taskAsync(new Runnable() {
-                @Override
-                public void run() {
-                    if (PS.this.IMP.initPlotMeConverter()) {
-                        PS.log("&c=== IMPORTANT ===");
-                        PS.log("&cTHIS MESSAGE MAY BE EXTREMELY HELPFUL IF YOU HAVE TROUBLE CONVERTING PlotMe!");
-                        PS.log("&c - Make sure 'UUID.read-from-disk' is disabled (false)!");
-                        PS.log("&c - Sometimes the database can be locked, deleting PlotMe.jar beforehand will fix the issue!");
-                        PS.log("&c - After the conversion is finished, please set 'plotme-converter' to false in the "
-                                + "'settings.yml'");
-                    }
-                    Settings.Enabled_Components.PLOTME_CONVERTER = false;
+            TaskManager.IMP.taskAsync(() -> {
+                if (PS.this.IMP.initPlotMeConverter()) {
+                    PS.log("&c=== IMPORTANT ===");
+                    PS.log("&cTHIS MESSAGE MAY BE EXTREMELY HELPFUL IF YOU HAVE TROUBLE CONVERTING PlotMe!");
+                    PS.log("&c - Make sure 'UUID.read-from-disk' is disabled (false)!");
+                    PS.log("&c - Sometimes the database can be locked, deleting PlotMe.jar beforehand will fix the issue!");
+                    PS.log("&c - After the conversion is finished, please set 'plotme-converter' to false in the "
+                            + "'settings.yml'");
                 }
+                Settings.Enabled_Components.PLOTME_CONVERTER = false;
             });
         }
     }
@@ -443,8 +404,6 @@ public class PS{
 
     /**
      * Check if `version` is >= `version2`.
-     * @param version
-     * @param version2
      * @return true if `version` is >= `version2`
      */
     public boolean checkVersion(int[] version, int... version2) {
@@ -476,7 +435,11 @@ public class PS{
 
     public PlotManager getPlotManager(Location location) {
         PlotArea pa = getPlotAreaAbs(location);
-        return pa != null ? pa.manager : null;
+        if (pa != null) {
+            return pa.manager;
+        } else {
+            return null;
+        }
     }
 
     /**
@@ -489,7 +452,11 @@ public class PS{
         HashMap<PlotId, Plot> plots;
         if (plots_tmp == null || (plots = plots_tmp.remove(plotArea.toString())) == null) {
             if (plotArea.TYPE == 2) {
-                plots = this.plots_tmp != null ? this.plots_tmp.get(plotArea.worldname) : null;
+                if (this.plots_tmp != null) {
+                    plots = this.plots_tmp.get(plotArea.worldname);
+                } else {
+                    plots = null;
+                }
                 if (plots != null) {
                     Iterator<Entry<PlotId, Plot>> iterator = plots.entrySet().iterator();
                     while (iterator.hasNext()) {
@@ -510,7 +477,11 @@ public class PS{
         Set<PlotCluster> clusters;
         if (clusters_tmp == null || (clusters = clusters_tmp.remove(plotArea.toString())) == null) {
             if (plotArea.TYPE == 2) {
-                clusters = this.clusters_tmp != null ? this.clusters_tmp.get(plotArea.worldname) : null;
+                if (this.clusters_tmp != null) {
+                    clusters = this.clusters_tmp.get(plotArea.worldname);
+                } else {
+                    clusters = null;
+                }
                 if (clusters != null) {
                     Iterator<PlotCluster> iterator = clusters.iterator();
                     while (iterator.hasNext()) {
@@ -553,11 +524,8 @@ public class PS{
         if (this.plots_tmp == null) {
             this.plots_tmp = new HashMap<>();
         }
-        HashMap<PlotId, Plot> map = this.plots_tmp.get(area.toString());
-        if (map == null) {
-            map = new HashMap<>();
-            this.plots_tmp.put(area.toString(), map);
-        }
+        HashMap<PlotId, Plot> map =
+            this.plots_tmp.computeIfAbsent(area.toString(), k -> new HashMap<>());
         for (Plot plot : area.getPlots()) {
             map.put(plot.getId(), plot);
         }
@@ -625,12 +593,7 @@ public class PS{
                 result.add(plot);
             }
         }
-        Collections.sort(overflow, new Comparator<Plot>() {
-            @Override
-            public int compare(Plot a, Plot b) {
-                return a.hashCode() - b.hashCode();
-            }
-        });
+        overflow.sort(Comparator.comparingInt(Plot::hashCode));
         result.addAll(overflow);
         return result;
     }
@@ -676,7 +639,7 @@ public class PS{
                 overflow.add(plot);
             }
         }
-        Plot[] overflowArray = overflow.toArray(new Plot[overflow.size()]);
+        Plot[] overflowArray = overflow.toArray(new Plot[0]);
         sortPlotsByHash(overflowArray);
         ArrayList<Plot> result = new ArrayList<>(cache.length + overflowArray.length);
         for (Plot plot : cache) {
@@ -685,9 +648,7 @@ public class PS{
             }
         }
         Collections.addAll(result, overflowArray);
-        for (Plot plot : extra) {
-            result.add(plot);
-        }
+        result.addAll(extra);
         return result;
     }
 
@@ -756,7 +717,7 @@ public class PS{
                 overflow.add(plot);
             }
         }
-        Plot[] overflowArray = overflow.toArray(new Plot[overflow.size()]);
+        Plot[] overflowArray = overflow.toArray(new Plot[0]);
         sortPlotsByHash(overflowArray);
         ArrayList<Plot> result = new ArrayList<>(cache.length + overflowArray.length);
         for (Plot plot : cache) {
@@ -765,15 +726,12 @@ public class PS{
             }
         }
         Collections.addAll(result, overflowArray);
-        for (Plot plot : extra) {
-            result.add(plot);
-        }
+        result.addAll(extra);
         return result;
     }
 
     /**
      * Sort plots by creation timestamp.
-     * @param input
      * @deprecated Unchecked, use {@link #sortPlots(Collection, SortType, PlotArea)} instead which will call this after checks
      * @return
      */
@@ -785,12 +743,7 @@ public class PS{
         } else {
             list = new ArrayList<>(input);
         }
-        Collections.sort(list, new Comparator<Plot>() {
-            @Override
-            public int compare(Plot a, Plot b) {
-                return Long.compare(ExpireManager.IMP.getTimestamp(a.owner), ExpireManager.IMP.getTimestamp(b.owner));
-            }
-        });
+        list.sort(Comparator.comparingLong(a -> ExpireManager.IMP.getTimestamp(a.owner)));
         return list;
     }
 
@@ -814,7 +767,7 @@ public class PS{
             }
         } else {
             for (PlotArea area : manager.getAllPlotAreas()) {
-                map.put(area, new ArrayList<Plot>(0));
+                map.put(area, new ArrayList<>(0));
             }
             Collection<Plot> lastList = null;
             PlotArea lastWorld = null;
@@ -829,18 +782,15 @@ public class PS{
             }
         }
         List<PlotArea> areas = Arrays.asList(manager.getAllPlotAreas());
-        Collections.sort(areas, new Comparator<PlotArea>() {
-            @Override
-            public int compare(PlotArea a, PlotArea b) {
-                if (priorityArea != null) {
-                    if (a.equals(priorityArea)) {
-                        return -1;
-                    } else if (b.equals(priorityArea)) {
-                        return 1;
-                    }
+        areas.sort((a, b) -> {
+            if (priorityArea != null) {
+                if (a.equals(priorityArea)) {
+                    return -1;
+                } else if (b.equals(priorityArea)) {
+                    return 1;
                 }
-                return a.hashCode() - b.hashCode();
             }
+            return a.hashCode() - b.hashCode();
         });
         ArrayList<Plot> toReturn = new ArrayList<>(plots.size());
         for (PlotArea area : areas) {
@@ -919,11 +869,8 @@ public class PS{
             String world = entry.getKey();
             PlotArea area = getPlotArea(world, null);
             if (area == null) {
-                HashMap<PlotId, Plot> map = this.plots_tmp.get(world);
-                if (map == null) {
-                    map = new HashMap<>();
-                    this.plots_tmp.put(world, map);
-                }
+                HashMap<PlotId, Plot> map =
+                    this.plots_tmp.computeIfAbsent(world, k -> new HashMap<>());
                 map.putAll(entry.getValue());
             } else {
                 for (Plot plot : entry.getValue().values()) {
@@ -1055,11 +1002,23 @@ public class PS{
     }
 
     public Collection<Plot> getPlots(PlotArea area) {
-        return area == null ? new HashSet<Plot>() : area.getPlots();
+        if (area == null) {
+            return new HashSet<>();
+        } else {
+            return area.getPlots();
+        }
     }
 
     public Plot getPlot(PlotArea area, PlotId id) {
-        return area == null ? null : id == null ? null : area.getPlot(id);
+        if (area == null) {
+            return null;
+        } else {
+            if (id == null) {
+                return null;
+            } else {
+                return area.getPlot(id);
+            }
+        }
     }
 
     public Set<Plot> getBasePlots(PlotPlayer player) {
@@ -1242,7 +1201,12 @@ public class PS{
                 PS.log(C.PREFIX + "&aDetected world load for '" + world + "'");
                 String gen_string = worldSection.getString("generator.plugin", IMP.getPluginName());
                 if (type == 2) {
-                    Set<PlotCluster> clusters = this.clusters_tmp != null ? this.clusters_tmp.get(world) : new HashSet<PlotCluster>();
+                    Set<PlotCluster> clusters;
+                    if (this.clusters_tmp != null) {
+                        clusters = this.clusters_tmp.get(world);
+                    } else {
+                        clusters = new HashSet<>();
+                    }
                     if (clusters == null) {
                         throw new IllegalArgumentException("No cluster exists for world: " + world);
                     }
@@ -1475,37 +1439,6 @@ public class PS{
 
     public File getJarFile() {
         return jarFile;
-    }
-
-    public boolean update(PlotPlayer sender, URL url) {
-        try {
-            String name = this.jarFile.getName();
-            File newJar = new File("plugins/update/" + name);
-            MainUtil.sendMessage(sender, "$1Downloading from provided URL: &7" + url);
-            URLConnection con = url.openConnection();
-            try (InputStream stream = con.getInputStream()) {
-                File parent = newJar.getParentFile();
-                if (!parent.exists()) {
-                    parent.mkdirs();
-                }
-                MainUtil.sendMessage(sender, "$2 - Output: " + newJar);
-                if (!newJar.delete()) {
-                    MainUtil.sendMessage(sender, "Failed to update " + IMP.getPluginName() + "");
-                    MainUtil.sendMessage(sender, "Jar file failed to delete.");
-                    MainUtil.sendMessage(sender, " - Please update manually");
-                }
-                Files.copy(stream, newJar.toPath());
-            }
-            MainUtil.sendMessage(sender, "$1The update will take effect when the server is restarted next");
-            return true;
-        } catch (IOException e) {
-            MainUtil.sendMessage(sender, "Failed to update " + IMP.getPluginName() + "");
-            MainUtil.sendMessage(sender, " - Please update manually");
-            PS.log("============ Stacktrace ============");
-            e.printStackTrace();
-            PS.log("====================================");
-        }
-        return false;
     }
 
     /**
@@ -1857,7 +1790,11 @@ public class PS{
 
     public PlotArea getFirstPlotArea() {
         PlotArea[] areas = manager.getAllPlotAreas();
-        return areas.length > 0 ? areas[0] : null;
+        if (areas.length > 0) {
+            return areas[0];
+        } else {
+            return null;
+        }
     }
 
     public int getPlotAreaCount() {
