@@ -3,6 +3,7 @@ package com.intellectualcrafters.plot.object;
 import com.google.common.base.Optional;
 import com.google.common.collect.BiMap;
 import com.google.common.collect.ImmutableSet;
+import com.google.common.collect.Sets;
 import com.intellectualcrafters.jnbt.CompoundTag;
 import com.intellectualcrafters.plot.PS;
 import com.intellectualcrafters.plot.config.C;
@@ -30,6 +31,7 @@ import com.intellectualcrafters.plot.util.expiry.ExpireManager;
 import com.intellectualcrafters.plot.util.expiry.PlotAnalysis;
 import com.plotsquared.listener.PlotListener;
 
+import javax.annotation.Nonnull;
 import java.awt.geom.Area;
 import java.awt.geom.PathIterator;
 import java.awt.geom.Rectangle2D;
@@ -51,18 +53,17 @@ import java.util.concurrent.ConcurrentHashMap;
 /**
  * The plot class<br>
  * [IMPORTANT]
- *  - Unclaimed plots will not have persistent information.
- *  - Any information set/modified in an unclaimed object may not be reflected in other instances
- *  - Using the `new` operator will create an unclaimed plot instance
- *  - Use the methods from the PlotArea/PS/Location etc to get existing plots
+ * - Unclaimed plots will not have persistent information.
+ * - Any information set/modified in an unclaimed object may not be reflected in other instances
+ * - Using the `new` operator will create an unclaimed plot instance
+ * - Use the methods from the PlotArea/PS/Location etc to get existing plots
  */
 public class Plot {
 
     /**
      * @deprecated raw access is deprecated
      */
-    @Deprecated
-    private static HashSet<Plot> connected_cache;
+    @Deprecated private static HashSet<Plot> connected_cache;
     private static HashSet<RegionWrapper> regions_cache;
     /**
      * The {@link PlotId}.
@@ -73,23 +74,22 @@ public class Plot {
      * (Merged plots can have multiple owners)
      * Direct access is Deprecated: use getOwners()
      */
-    @Deprecated
-    public UUID owner;
+    @Deprecated public UUID owner;
     /**
      * Has the plot changed since the last save cycle?
      */
     public boolean countsTowardsMax = true;
     /**
      * Represents whatever the database manager needs it to: <br>
-     *  - A value of -1 usually indicates the plot will not be stored in the DB<br>
-     *  - A value of 0 usually indicates that the DB manager hasn't set a value<br>
+     * - A value of -1 usually indicates the plot will not be stored in the DB<br>
+     * - A value of 0 usually indicates that the DB manager hasn't set a value<br>
+     *
      * @deprecated magical
      */
-    @Deprecated
-    public int temp;
+    @Deprecated public int temp;
     /**
      * Plot creation timestamp (not accurate if the plot was created before this was implemented)<br>
-     *  - Milliseconds since the epoch<br>
+     * - Milliseconds since the epoch<br>
      */
     private long timestamp;
     /**
@@ -106,8 +106,8 @@ public class Plot {
     private HashSet<UUID> denied;
     /**
      * External settings class.
-     *  - Please favor the methods over direct access to this class<br>
-     *  - The methods are more likely to be left unchanged from version changes<br>
+     * - Please favor the methods over direct access to this class<br>
+     * - The methods are more likely to be left unchanged from version changes<br>
      */
     private PlotSettings settings;
     /**
@@ -117,13 +117,14 @@ public class Plot {
     /**
      * Session only plot metadata (session is until the server stops)<br>
      * <br>
-     *  For persistent metadata use the flag system
-     *  @see FlagManager
+     * For persistent metadata use the flag system
+     *
+     * @see FlagManager
      */
     private ConcurrentHashMap<String, Object> meta;
     /**
      * The cached origin plot.
-     *  - The origin plot is used for plot grouping and relational data
+     * - The origin plot is used for plot grouping and relational data
      */
     private Plot origin;
 
@@ -131,11 +132,10 @@ public class Plot {
      * Constructor for a new plot.
      * (Only changes after plot.create() will be properly set in the database)
      *
-     * @see Plot#getPlot(Location) for existing plots
-     *
-     * @param area the PlotArea where the plot is located
-     * @param id the plot id
+     * @param area  the PlotArea where the plot is located
+     * @param id    the plot id
      * @param owner the plot owner
+     * @see Plot#getPlot(Location) for existing plots
      */
     public Plot(PlotArea area, PlotId id, UUID owner) {
         this.area = area;
@@ -147,10 +147,9 @@ public class Plot {
      * Constructor for an unowned plot.
      * (Only changes after plot.create() will be properly set in the database)
      *
-     * @see Plot#getPlot(Location) for existing plots
-     *
      * @param area the PlotArea where the plot is located
-     * @param id the plot id
+     * @param id   the plot id
+     * @see Plot#getPlot(Location) for existing plots
      */
     public Plot(PlotArea area, PlotId id) {
         this.area = area;
@@ -162,12 +161,11 @@ public class Plot {
      * The database will ignore any queries regarding temporary plots.
      * Please note that some bulk plot management functions may still affect temporary plots (TODO: fix this)
      *
-     * @see Plot#getPlot(Location) for existing plots
-     *
-     * @param area the PlotArea where the plot is located
-     * @param id the plot id
+     * @param area  the PlotArea where the plot is located
+     * @param id    the plot id
      * @param owner the owner of the plot
      * @param temp
+     * @see Plot#getPlot(Location) for existing plots
      */
     public Plot(PlotArea area, PlotId id, UUID owner, int temp) {
         this.area = area;
@@ -179,17 +177,16 @@ public class Plot {
     /**
      * Constructor for a saved plots (Used by the database manager when plots are fetched)
      *
-     * @see Plot#getPlot(Location) for existing plots
-     *
-     * @param id the plot id
-     * @param owner the plot owner
+     * @param id      the plot id
+     * @param owner   the plot owner
      * @param trusted
      * @param denied
      * @param merged
+     * @see Plot#getPlot(Location) for existing plots
      */
-    public Plot(PlotId id, UUID owner, HashSet<UUID> trusted, HashSet<UUID> members, HashSet<UUID> denied,
-            String alias, BlockLoc position,
-            Collection<Flag> flags, PlotArea area, boolean[] merged, long timestamp, int temp) {
+    public Plot(PlotId id, UUID owner, HashSet<UUID> trusted, HashSet<UUID> members,
+        HashSet<UUID> denied, String alias, BlockLoc position, Collection<Flag> flags,
+        PlotArea area, boolean[] merged, long timestamp, int temp) {
         this.id = id;
         this.area = area;
         this.owner = owner;
@@ -215,8 +212,9 @@ public class Plot {
 
     /**
      * Get a plot from a string e.g. [area];[id]
+     *
      * @param defaultArea If no area is specified
-     * @param string plot id/area + id
+     * @param string      plot id/area + id
      * @return New or existing plot object
      */
     public static Plot fromString(PlotArea defaultArea, String string) {
@@ -224,7 +222,11 @@ public class Plot {
         if (split.length == 2) {
             if (defaultArea != null) {
                 PlotId id = PlotId.fromString(split[0] + ';' + split[1]);
-                return id != null ? defaultArea.getPlotAbs(id) : null;
+                if (id != null) {
+                    return defaultArea.getPlotAbs(id);
+                } else {
+                    return null;
+                }
             }
         } else if (split.length == 3) {
             PlotArea pa = PS.get().getPlotArea(split[0], null);
@@ -245,10 +247,9 @@ public class Plot {
     /**
      * Return a new/cached plot object at a given location.
      *
-     * @see PlotPlayer#getCurrentPlot() if a player is expected here.
-     *
      * @param location the location of the plot
      * @return
+     * @see PlotPlayer#getCurrentPlot() if a player is expected here.
      */
     public static Plot getPlot(Location location) {
         PlotArea pa = PS.get().getPlotAreaAbs(location);
@@ -262,9 +263,10 @@ public class Plot {
      * Session only plot metadata (session is until the server stops)<br>
      * <br>
      * For persistent metadata use the flag system
-     * @see FlagManager
+     *
      * @param key
      * @param value
+     * @see FlagManager
      */
     public void setMeta(String key, Object value) {
         if (this.meta == null) {
@@ -277,6 +279,7 @@ public class Plot {
      * Get the metadata for a key<br>
      * <br>
      * For persistent metadata use the flag system
+     *
      * @param key
      * @return
      */
@@ -289,8 +292,9 @@ public class Plot {
 
     /**
      * Delete the metadata for a key<br>
-     *  - metadata is session only
-     *  - deleting other plugin's metadata may cause issues
+     * - metadata is session only
+     * - deleting other plugin's metadata may cause issues
+     *
      * @param key
      */
     public void deleteMeta(String key) {
@@ -301,6 +305,7 @@ public class Plot {
 
     /**
      * Get the cluster this plot is associated with
+     *
      * @return the PlotCluster object, or null
      */
     public PlotCluster getCluster() {
@@ -309,8 +314,9 @@ public class Plot {
 
     /**
      * Efficiently get the players currently inside this plot<br>
-     *  - Will return an empty list if no players are in the plot<br>
-     *  - Remember, you can cast a PlotPlayer to it's respective implementation (BukkitPlayer, SpongePlayer) to obtain the player object
+     * - Will return an empty list if no players are in the plot<br>
+     * - Remember, you can cast a PlotPlayer to it's respective implementation (BukkitPlayer, SpongePlayer) to obtain the player object
+     *
      * @return list of PlotPlayer(s) or an empty list
      */
     public List<PlotPlayer> getPlayersInPlot() {
@@ -335,6 +341,7 @@ public class Plot {
 
     /**
      * Check if a UUID is a plot owner (merged plots may have multiple owners)
+     *
      * @param uuid the player uuid
      * @return if the provided uuid is the owner of the plot
      */
@@ -360,6 +367,7 @@ public class Plot {
 
     /**
      * Get a immutable set of owner UUIDs for a plot (supports multi-owner mega-plots).
+     *
      * @return the plot owners
      */
     public Set<UUID> getOwners() {
@@ -373,7 +381,8 @@ public class Plot {
             UUID last = this.owner;
             owners.add(this.owner);
             for (Plot current : array) {
-                if (last == null || current.owner.getMostSignificantBits() != last.getMostSignificantBits()) {
+                if (last == null || current.owner.getMostSignificantBits() != last
+                    .getMostSignificantBits()) {
                     owners.add(current.owner);
                     last = current.owner;
                 }
@@ -387,7 +396,6 @@ public class Plot {
      * Check if the player is either the owner or on the trusted/added list.
      *
      * @param uuid
-     *
      * @return true if the player is added/trusted or is the owner
      */
     public boolean isAdded(UUID uuid) {
@@ -413,12 +421,11 @@ public class Plot {
      * Should the player be denied from entering.
      *
      * @param uuid
-     *
      * @return boolean false if the player is allowed to enter
      */
     public boolean isDenied(UUID uuid) {
         return this.denied != null && (this.denied.contains(DBFunc.everyone) && !this.isAdded(uuid)
-                || !this.isAdded(uuid) && this.denied.contains(uuid));
+            || !this.isAdded(uuid) && this.denied.contains(uuid));
     }
 
     /**
@@ -430,7 +437,8 @@ public class Plot {
 
     /**
      * Get the plot world object for this plot<br>
-     *  - The generic PlotArea object can be casted to its respective class for more control (e.g. HybridPlotWorld)
+     * - The generic PlotArea object can be casted to its respective class for more control (e.g. HybridPlotWorld)
+     *
      * @return PlotArea
      */
     public PlotArea getArea() {
@@ -441,6 +449,7 @@ public class Plot {
      * Assign this plot to a plot area.<br>
      * (Mostly used during startup when worlds are being created)<br>
      * Note: Using this when it doesn't make sense will result in strange behavior
+     *
      * @param area
      */
     public void setArea(PlotArea area) {
@@ -456,7 +465,8 @@ public class Plot {
 
     /**
      * Get the plot manager object for this plot<br>
-     *  - The generic PlotManager object can be casted to its respective class for more control (e.g. HybridPlotManager)
+     * - The generic PlotManager object can be casted to its respective class for more control (e.g. HybridPlotManager)
+     *
      * @return PlotManager
      */
     public PlotManager getManager() {
@@ -465,11 +475,11 @@ public class Plot {
 
     /**
      * Get or create plot settings.
+     *
      * @return PlotSettings
      * @deprecated use equivalent plot method;
      */
-    @Deprecated
-    public PlotSettings getSettings() {
+    @Deprecated public PlotSettings getSettings() {
         if (this.settings == null) {
             this.settings = new PlotSettings();
         }
@@ -479,6 +489,7 @@ public class Plot {
     /**
      * Returns true if the plot is not merged, or it is the base
      * plot of multiple merged plots.
+     *
      * @return
      */
     public boolean isBasePlot() {
@@ -487,10 +498,11 @@ public class Plot {
 
     /**
      * The base plot is an arbitrary but specific connected plot. It is useful for the following:<br>
-     *  - Merged plots need to be treated as a single plot for most purposes<br>
-     *  - Some data such as home location needs to be associated with the group rather than each plot<br>
-     *  - If the plot is not merged it will return itself.<br>
-     *  - The result is cached locally
+     * - Merged plots need to be treated as a single plot for most purposes<br>
+     * - Some data such as home location needs to be associated with the group rather than each plot<br>
+     * - If the plot is not merged it will return itself.<br>
+     * - The result is cached locally
+     *
      * @return base Plot
      */
     public Plot getBasePlot(boolean recalculate) {
@@ -520,16 +532,19 @@ public class Plot {
 
     /**
      * Check if the plot is merged in any direction.
+     *
      * @return is the plot merged or not
      */
     public boolean isMerged() {
-        return getSettings().getMerged(0) || getSettings().getMerged(2) || getSettings().getMerged(1) || getSettings().getMerged(3);
+        return getSettings().getMerged(0) || getSettings().getMerged(2) || getSettings()
+            .getMerged(1) || getSettings().getMerged(3);
     }
 
     /**
      * Get the timestamp of when the plot was created (unreliable)<br>
      * - not accurate if the plot was created before this was implemented<br>
-     *  - Milliseconds since the epoch<br>
+     * - Milliseconds since the epoch<br>
+     *
      * @return the creation date of the plot
      */
     public long getTimestamp() {
@@ -553,6 +568,7 @@ public class Plot {
      * 7 = north-west<br>
      * ----------<br>
      * Note: A plot that is merged north and east will not be merged northeast if the northeast plot is not part of the same group<br>
+     *
      * @param direction
      * @return true if merged in that direction
      */
@@ -584,10 +600,9 @@ public class Plot {
             case 6:
                 i = direction - 4;
                 i2 = direction - 3;
-                return this.getSettings().getMerged(i2)
-                        && this.getSettings().getMerged(i)
-                        && this.area.getPlotAbs(this.id.getRelative(i)).getMerged(i2)
-                        && this.area.getPlotAbs(this.id.getRelative(i2)).getMerged(i);
+                return this.getSettings().getMerged(i2) && this.getSettings().getMerged(i)
+                    && this.area.getPlotAbs(this.id.getRelative(i)).getMerged(i2) && this.area
+                    .getPlotAbs(this.id.getRelative(i2)).getMerged(i);
 
         }
         return false;
@@ -595,6 +610,7 @@ public class Plot {
 
     /**
      * Get the denied users.
+     *
      * @return a set of denied users
      */
     public HashSet<UUID> getDenied() {
@@ -606,12 +622,22 @@ public class Plot {
 
     /**
      * Set the denied users for this plot.
+     *
      * @param uuids
      */
     public void setDenied(Set<UUID> uuids) {
         boolean larger = uuids.size() > getDenied().size();
-        HashSet<UUID> intersection = new HashSet<>(larger ? getDenied() : uuids);
-        intersection.retainAll(larger ? uuids : getDenied());
+        HashSet<UUID> intersection;
+        if (larger) {
+            intersection = new HashSet<>(getDenied());
+        } else {
+            intersection = new HashSet<>(uuids);
+        }
+        if (larger) {
+            intersection.retainAll(uuids);
+        } else {
+            intersection.retainAll(getDenied());
+        }
         uuids.removeAll(intersection);
         HashSet<UUID> toRemove = new HashSet<>(getDenied());
         toRemove.removeAll(intersection);
@@ -625,6 +651,7 @@ public class Plot {
 
     /**
      * Get the trusted users.
+     *
      * @return a set of trusted users
      */
     public HashSet<UUID> getTrusted() {
@@ -636,12 +663,22 @@ public class Plot {
 
     /**
      * Set the trusted users for this plot.
+     *
      * @param uuids
      */
     public void setTrusted(Set<UUID> uuids) {
         boolean larger = uuids.size() > getTrusted().size();
-        HashSet<UUID> intersection = new HashSet<>(larger ? getTrusted() : uuids);
-        intersection.retainAll(larger ? uuids : getTrusted());
+        HashSet<UUID> intersection;
+        if (larger) {
+            intersection = new HashSet<>(getTrusted());
+        } else {
+            intersection = new HashSet<>(uuids);
+        }
+        if (larger) {
+            intersection.retainAll(uuids);
+        } else {
+            intersection.retainAll(getTrusted());
+        }
         uuids.removeAll(intersection);
         HashSet<UUID> toRemove = new HashSet<>(getTrusted());
         toRemove.removeAll(intersection);
@@ -655,6 +692,7 @@ public class Plot {
 
     /**
      * Get the members
+     *
      * @return a set of members
      */
     public HashSet<UUID> getMembers() {
@@ -666,12 +704,22 @@ public class Plot {
 
     /**
      * Set the members for this plot
+     *
      * @param uuids
      */
     public void setMembers(Set<UUID> uuids) {
         boolean larger = uuids.size() > getMembers().size();
-        HashSet<UUID> intersection = new HashSet<>(larger ? getMembers() : uuids);
-        intersection.retainAll(larger ? uuids : getMembers());
+        HashSet<UUID> intersection;
+        if (larger) {
+            intersection = new HashSet<>(getMembers());
+        } else {
+            intersection = new HashSet<>(uuids);
+        }
+        if (larger) {
+            intersection.retainAll(uuids);
+        } else {
+            intersection.retainAll(getMembers());
+        }
         uuids.removeAll(intersection);
         HashSet<UUID> toRemove = new HashSet<>(getMembers());
         toRemove.removeAll(intersection);
@@ -685,6 +733,7 @@ public class Plot {
 
     /**
      * Deny someone (updates database as well)
+     *
      * @param uuid the uuid of the player to deny.
      */
     public void addDenied(UUID uuid) {
@@ -723,6 +772,7 @@ public class Plot {
 
     /**
      * Set the plot owner (and update the database)
+     *
      * @param owner
      */
     public void setOwner(UUID owner) {
@@ -748,14 +798,22 @@ public class Plot {
 
     /**
      * Set the plot owner (and update the database)
+     *
      * @param owner
      * @param initiator
      * @return boolean
      */
     public boolean setOwner(UUID owner, PlotPlayer initiator) {
-        boolean result = EventUtil.manager.callOwnerChange(initiator, this, owner, hasOwner() ? this.owner : null, hasOwner());
-        if(!result)
+        boolean result;
+        if (hasOwner()) {
+            result =
+                EventUtil.manager.callOwnerChange(initiator, this, owner, this.owner, hasOwner());
+        } else {
+            result = EventUtil.manager.callOwnerChange(initiator, this, owner, null, hasOwner());
+        }
+        if (!result) {
             return false;
+        }
         if (!hasOwner()) {
             this.owner = owner;
             create();
@@ -779,9 +837,10 @@ public class Plot {
 
     /**
      * Clear a plot.
+     *
+     * @param whenDone A runnable to execute when clearing finishes, or null
      * @see this#clear(boolean, boolean, Runnable)
      * @see #deletePlot(Runnable) to clear and delete a plot
-     * @param whenDone A runnable to execute when clearing finishes, or null
      */
     public void clear(Runnable whenDone) {
         this.clear(false, false, whenDone);
@@ -792,11 +851,11 @@ public class Plot {
             return false;
         }
         if (isDelete) {
-            if(!EventUtil.manager.callDelete(this)) {
+            if (!EventUtil.manager.callDelete(this)) {
                 return false;
             }
         } else {
-            if(!EventUtil.manager.callClear(this)) {
+            if (!EventUtil.manager.callClear(this)) {
                 return false;
             }
         }
@@ -809,12 +868,10 @@ public class Plot {
         this.unlinkPlot(true, !isDelete);
         final PlotManager manager = this.area.getPlotManager();
         Runnable run = new Runnable() {
-            @Override
-            public void run() {
+            @Override public void run() {
                 if (queue.isEmpty()) {
                     Runnable run = new Runnable() {
-                        @Override
-                        public void run() {
+                        @Override public void run() {
                             for (RegionWrapper region : regions) {
                                 Location[] corners = region.getCorners(getWorldName());
                                 ChunkManager.manager.clearAllEntities(corners[0], corners[1]);
@@ -834,7 +891,8 @@ public class Plot {
                 }
                 Plot current = queue.poll();
                 if (Plot.this.area.TERRAIN != 0) {
-                    ChunkManager.manager.regenerateRegion(current.getBottomAbs(), current.getTopAbs(), false, this);
+                    ChunkManager.manager
+                        .regenerateRegion(current.getBottomAbs(), current.getTopAbs(), false, this);
                     return;
                 }
                 manager.clearPlot(Plot.this.area, current, this);
@@ -846,34 +904,40 @@ public class Plot {
 
     /**
      * Set the biome for a plot asynchronously
-     * @param biome The biome e.g. "forest"
+     *
+     * @param biome    The biome e.g. "forest"
      * @param whenDone The task to run when finished, or null
      */
     public void setBiome(final String biome, final Runnable whenDone) {
         final ArrayDeque<RegionWrapper> regions = new ArrayDeque<>(this.getRegions());
         final int extendBiome;
         if (area instanceof SquarePlotWorld) {
-            extendBiome = (((SquarePlotWorld) area).ROAD_WIDTH > 0) ? 1 : 0;
+            if (((SquarePlotWorld) area).ROAD_WIDTH > 0) {
+                extendBiome = 1;
+            } else {
+                extendBiome = 0;
+            }
         } else {
             extendBiome = 0;
         }
         Runnable run = new Runnable() {
-            @Override
-            public void run() {
+            @Override public void run() {
                 if (regions.isEmpty()) {
                     Plot.this.refreshChunks();
                     TaskManager.runTask(whenDone);
                     return;
                 }
                 RegionWrapper region = regions.poll();
-                Location pos1 = new Location(getWorldName(), region.minX - extendBiome, region.minY, region.minZ - extendBiome);
-                Location pos2 = new Location(getWorldName(), region.maxX + extendBiome, region.maxY, region.maxZ + extendBiome);
+                Location pos1 = new Location(getWorldName(), region.minX - extendBiome, region.minY,
+                    region.minZ - extendBiome);
+                Location pos2 = new Location(getWorldName(), region.maxX + extendBiome, region.maxY,
+                    region.maxZ + extendBiome);
                 ChunkManager.chunkTask(pos1, pos2, new RunnableVal<int[]>() {
-                    @Override
-                    public void run(int[] value) {
+                    @Override public void run(int[] value) {
                         ChunkLoc loc = new ChunkLoc(value[0], value[1]);
                         ChunkManager.manager.loadChunk(getWorldName(), loc, false);
-                        MainUtil.setBiome(getWorldName(), value[2], value[3], value[4], value[5], biome);
+                        MainUtil.setBiome(getWorldName(), value[2], value[3], value[4], value[5],
+                            biome);
                         ChunkManager.manager.unloadChunk(getWorldName(), loc, true, true);
                     }
                 }, this, 5);
@@ -885,6 +949,7 @@ public class Plot {
 
     /**
      * Unlink the plot and all connected plots.
+     *
      * @param createSign
      * @param createRoad
      * @return
@@ -927,13 +992,12 @@ public class Plot {
             }
         }
         for (Plot current : plots) {
-            boolean[] merged = new boolean[]{false, false, false, false};
+            boolean[] merged = new boolean[] {false, false, false, false};
             current.setMerged(merged);
         }
         if (createSign) {
             GlobalBlockQueue.IMP.addTask(new Runnable() {
-                @Override
-                public void run() {
+                @Override public void run() {
                     for (Plot current : plots) {
                         current.setSign(MainUtil.getName(current.owner));
                     }
@@ -948,28 +1012,25 @@ public class Plot {
 
     /**
      * Set the sign for a plot to a specific name
+     *
      * @param name
      */
     public void setSign(final String name) {
-        if (!isLoaded()) return;
+        if (!isLoaded()) {
+            return;
+        }
         if (!PS.get().isMainThread(Thread.currentThread())) {
-            TaskManager.runTask(new Runnable() {
-                @Override
-                public void run() {
-                    Plot.this.setSign(name);
-                }
-            });
+            TaskManager.runTask(() -> Plot.this.setSign(name));
             return;
         }
         PlotManager manager = this.area.getPlotManager();
         if (this.area.ALLOW_SIGNS) {
             Location loc = manager.getSignLoc(this.area, this);
             String id = this.id.x + ";" + this.id.y;
-            String[] lines = new String[]{
-                    C.OWNER_SIGN_LINE_1.formatted().replaceAll("%id%", id),
-                    C.OWNER_SIGN_LINE_2.formatted().replaceAll("%id%", id).replaceAll("%plr%", name),
-                    C.OWNER_SIGN_LINE_3.formatted().replaceAll("%id%", id).replaceAll("%plr%", name),
-                    C.OWNER_SIGN_LINE_4.formatted().replaceAll("%id%", id).replaceAll("%plr%", name)};
+            String[] lines = new String[] {C.OWNER_SIGN_LINE_1.formatted().replaceAll("%id%", id),
+                C.OWNER_SIGN_LINE_2.formatted().replaceAll("%id%", id).replaceAll("%plr%", name),
+                C.OWNER_SIGN_LINE_3.formatted().replaceAll("%id%", id).replaceAll("%plr%", name),
+                C.OWNER_SIGN_LINE_4.formatted().replaceAll("%id%", id).replaceAll("%plr%", name)};
             WorldUtil.IMP.setSign(this.getWorldName(), loc.getX(), loc.getY(), loc.getZ(), lines);
         }
     }
@@ -980,6 +1041,7 @@ public class Plot {
 
     /**
      * This will return null if the plot hasn't been analyzed
+     *
      * @return analysis of plot
      */
     public PlotAnalysis getComplexity(Settings.Auto_Clear settings) {
@@ -992,6 +1054,7 @@ public class Plot {
 
     /**
      * Set a flag for this plot
+     *
      * @param flag
      * @param value
      */
@@ -1004,6 +1067,7 @@ public class Plot {
 
     /**
      * Remove a flag from this plot
+     *
      * @param flag the flag to remove
      * @return
      */
@@ -1013,6 +1077,7 @@ public class Plot {
 
     /**
      * Get the flag for a given key
+     *
      * @param key
      */
     public <V> Optional<V> getFlag(Flag<V> key) {
@@ -1021,7 +1086,8 @@ public class Plot {
 
     /**
      * Get the flag for a given key
-     * @param key the flag
+     *
+     * @param key          the flag
      * @param defaultValue if the key is null, the value to return
      */
     public <V> V getFlag(Flag<V> key, V defaultValue) {
@@ -1035,6 +1101,7 @@ public class Plot {
 
     /**
      * Delete a plot (use null for the runnable if you don't need to be notified on completion)
+     *
      * @see PS#removePlot(Plot, boolean)
      * @see #clear(Runnable) to simply clear a plot
      */
@@ -1044,8 +1111,7 @@ public class Plot {
         }
         final Set<Plot> plots = this.getConnectedPlots();
         this.clear(false, true, new Runnable() {
-            @Override
-            public void run() {
+            @Override public void run() {
                 for (Plot current : plots) {
                     current.unclaim();
                 }
@@ -1057,6 +1123,8 @@ public class Plot {
 
     /**
      * Count the entities in a plot
+     *
+     * @return
      * @see ChunkManager#countEntities(Plot)
      * 0 = Entity
      * 1 = Animal
@@ -1064,7 +1132,6 @@ public class Plot {
      * 3 = Mob
      * 4 = Boat
      * 5 = Misc
-     * @return
      */
     public int[] countEntities() {
         int[] count = new int[6];
@@ -1082,6 +1149,7 @@ public class Plot {
 
     /**
      * Returns true if a previous task was running
+     *
      * @return true if a previous task is running
      */
     public int addRunning() {
@@ -1094,7 +1162,8 @@ public class Plot {
 
     /**
      * Decrement the number of tracked tasks this plot is running<br>
-     *     - Used to track/limit the number of things a player can do on the plot at once
+     * - Used to track/limit the number of things a player can do on the plot at once
+     *
      * @return previous number of tasks (int)
      */
     public int removeRunning() {
@@ -1113,16 +1182,22 @@ public class Plot {
 
     /**
      * Get the number of tracked running tasks for this plot<br>
-     *     - Used to track/limit the number of things a player can do on the plot at once
+     * - Used to track/limit the number of things a player can do on the plot at once
+     *
      * @return number of tasks (int)
      */
     public int getRunning() {
         Integer value = (Integer) this.getMeta("running");
-        return value == null ? 0 : value;
+        if (value == null) {
+            return 0;
+        } else {
+            return value;
+        }
     }
 
     /**
      * Unclaim the plot (does not modify terrain). Changes made to this plot will not be reflected in unclaimed plot objects.
+     *
      * @return false if the Plot has no owner, otherwise true.
      */
     public boolean unclaim() {
@@ -1147,8 +1222,9 @@ public class Plot {
 
     /**
      * Unlink a plot and remove the roads
-     * @see this#unlinkPlot(boolean, boolean)
+     *
      * @return true if plot was linked
+     * @see this#unlinkPlot(boolean, boolean)
      */
     public boolean unlink() {
         return this.unlinkPlot(true, true);
@@ -1158,14 +1234,18 @@ public class Plot {
         Location[] corners = getCorners();
         Location top = corners[0];
         Location bot = corners[1];
-        Location loc = new Location(this.getWorldName(), MathMan.average(bot.getX(), top.getX()), MathMan.average(bot.getY(), top.getY()), MathMan.average(bot.getZ(), top.getZ()));
-        if (!isLoaded()) return loc;
-        int y = isLoaded() ? WorldUtil.IMP.getHighestBlock(getWorldName(), loc.getX(), loc.getZ()) : 64;
-        if (area.ALLOW_SIGNS) {
-            y = Math.max(y, getManager().getSignLoc(area, this).getY());
+        Location loc = new Location(this.getWorldName(), MathMan.average(bot.getX(), top.getX()),
+            MathMan.average(bot.getY(), top.getY()), MathMan.average(bot.getZ(), top.getZ()));
+        if (isLoaded()) {
+            int y = WorldUtil.IMP.getHighestBlock(getWorldName(), loc.getX(), loc.getZ());
+            if (area.ALLOW_SIGNS) {
+                y = Math.max(y, getManager().getSignLoc(area, this).getY());
+            }
+            loc.setY(1 + y);
+            return loc;
+        } else {
+            return loc;
         }
-        loc.setY(1 + y);
-        return loc;
     }
 
     public Location getSide() {
@@ -1173,7 +1253,12 @@ public class Plot {
         int x = (largest.maxX >> 1) - (largest.minX >> 1) + largest.minX;
         int z = largest.minZ - 1;
         PlotManager manager = getManager();
-        int y = isLoaded() ? WorldUtil.IMP.getHighestBlock(getWorldName(), x, z) : 64;
+        int y;
+        if (isLoaded()) {
+            y = WorldUtil.IMP.getHighestBlock(getWorldName(), x, z);
+        } else {
+            y = 64;
+        }
         if (area.ALLOW_SIGNS && (y <= 0 || y >= 255)) {
             y = Math.max(y, manager.getSignLoc(area, this).getY() - 1);
         }
@@ -1182,6 +1267,7 @@ public class Plot {
 
     /**
      * Return the home location for the plot
+     *
      * @return Home location
      */
     public Location getHome() {
@@ -1190,10 +1276,15 @@ public class Plot {
             return this.getDefaultHome(true);
         } else {
             Location bot = this.getBottomAbs();
-            Location loc = new Location(bot.getWorld(), bot.getX() + home.x, bot.getY() + home.y, bot.getZ() + home.z, home.yaw, home.pitch);
-            if (!isLoaded()) return loc;
+            Location loc = new Location(bot.getWorld(), bot.getX() + home.x, bot.getY() + home.y,
+                bot.getZ() + home.z, home.yaw, home.pitch);
+            if (!isLoaded()) {
+                return loc;
+            }
             if (WorldUtil.IMP.getBlock(loc).id != 0) {
-                loc.setY(Math.max(1 + WorldUtil.IMP.getHighestBlock(this.getWorldName(), loc.getX(), loc.getZ()), bot.getY()));
+                loc.setY(Math.max(
+                    1 + WorldUtil.IMP.getHighestBlock(this.getWorldName(), loc.getX(), loc.getZ()),
+                    bot.getY()));
             }
             return loc;
         }
@@ -1201,6 +1292,7 @@ public class Plot {
 
     /**
      * Set the home location
+     *
      * @param location
      */
     public void setHome(BlockLoc location) {
@@ -1218,7 +1310,8 @@ public class Plot {
 
     /**
      * Get the default home location for a plot<br>
-     *  - Ignores any home location set for that specific plot
+     * - Ignores any home location set for that specific plot
+     *
      * @return
      */
     public Location getDefaultHome() {
@@ -1227,7 +1320,12 @@ public class Plot {
 
     public Location getDefaultHome(boolean member) {
         Plot plot = this.getBasePlot(false);
-        PlotLoc loc = member ? area.DEFAULT_HOME : area.NONMEMBER_HOME;
+        PlotLoc loc;
+        if (member) {
+            loc = area.DEFAULT_HOME;
+        } else {
+            loc = area.NONMEMBER_HOME;
+        }
         if (loc != null) {
             int x;
             int z;
@@ -1242,7 +1340,12 @@ public class Plot {
                 x = bot.getX() + loc.x;
                 z = bot.getZ() + loc.z;
             }
-            int y = loc.y < 1 ? (isLoaded() ? WorldUtil.IMP.getHighestBlock(plot.getWorldName(), x, z) + 1 : 63) : loc.y;
+            int y;
+            if (loc.y < 1) {
+                y = isLoaded() ? WorldUtil.IMP.getHighestBlock(plot.getWorldName(), x, z) + 1 : 63;
+            } else {
+                y = loc.y;
+            }
             return new Location(plot.getWorldName(), x, y, z);
         }
         // Side
@@ -1252,13 +1355,16 @@ public class Plot {
     public double getVolume() {
         double count = 0;
         for (RegionWrapper region : getRegions()) {
-            count += (region.maxX - (double) region.minX + 1) * (region.maxZ - (double) region.minZ + 1) * 256;
+            count +=
+                (region.maxX - (double) region.minX + 1) * (region.maxZ - (double) region.minZ + 1)
+                    * 256;
         }
         return count;
     }
 
     /**
      * Get the average rating of the plot. This is the value displayed in /plot info
+     *
      * @return average rating as double
      */
     public double getAverageRating() {
@@ -1272,7 +1378,8 @@ public class Plot {
 
     /**
      * Set a rating for a user<br>
-     *  - If the user has already rated, the following will return false
+     * - If the user has already rated, the following will return false
+     *
      * @param uuid
      * @param rating
      * @return
@@ -1289,7 +1396,9 @@ public class Plot {
         return true;
     }
 
-    /** Clear the ratings for this plot */
+    /**
+     * Clear the ratings for this plot
+     */
     public void clearRatings() {
         Plot base = this.getBasePlot(false);
         PlotSettings baseSettings = base.getSettings();
@@ -1301,7 +1410,8 @@ public class Plot {
 
     /**
      * Get the ratings associated with a plot<br>
-     *  - The rating object may contain multiple categories
+     * - The rating object may contain multiple categories
+     *
      * @return Map of user who rated to the rating
      */
     public HashMap<UUID, Rating> getRatings() {
@@ -1339,7 +1449,9 @@ public class Plot {
         }
     }
 
-    /** Remove the plot sign if it is set. */
+    /**
+     * Remove the plot sign if it is set.
+     */
     public void removeSign() {
         PlotManager manager = this.area.getPlotManager();
         if (!this.area.ALLOW_SIGNS) {
@@ -1351,7 +1463,9 @@ public class Plot {
         queue.flush();
     }
 
-    /** Set the plot sign if plot signs are enabled. */
+    /**
+     * Set the plot sign if plot signs are enabled.
+     */
     public void setSign() {
         if (this.owner == null) {
             this.setSign("unknown");
@@ -1362,9 +1476,10 @@ public class Plot {
 
     /**
      * Register a plot and create it in the database<br>
-     *  - The plot will not be created if the owner is null<br>
-     *  - Any setting from before plot creation will not be saved until the server is stopped properly. i.e. Set any values/options after plot
-     *  creation.
+     * - The plot will not be created if the owner is null<br>
+     * - Any setting from before plot creation will not be saved until the server is stopped properly. i.e. Set any values/options after plot
+     * creation.
+     *
      * @return true if plot was created successfully
      */
     public boolean create() {
@@ -1378,7 +1493,8 @@ public class Plot {
         return claim(player, teleport, schematic, true);
     }
 
-    public boolean claim(final PlotPlayer player, boolean teleport, String schematic, boolean updateDB) {
+    public boolean claim(final PlotPlayer player, boolean teleport, String schematic,
+        boolean updateDB) {
         boolean result = EventUtil.manager.callClaim(player, this, false);
         if (updateDB) {
             if (!result || (!create(player.getUUID(), true))) {
@@ -1403,16 +1519,16 @@ public class Plot {
                     sch = SchematicHandler.manager.getSchematic(plotworld.SCHEMATIC_FILE);
                 }
             }
-            SchematicHandler.manager.paste(sch, this, 0, 0, 0, Settings.Schematics.PASTE_ON_TOP, new RunnableVal<Boolean>() {
-                @Override
-                public void run(Boolean value) {
-                    if (value) {
-                        MainUtil.sendMessage(player, C.SCHEMATIC_PASTE_SUCCESS);
-                    } else {
-                        MainUtil.sendMessage(player, C.SCHEMATIC_PASTE_FAILED);
+            SchematicHandler.manager.paste(sch, this, 0, 0, 0, Settings.Schematics.PASTE_ON_TOP,
+                new RunnableVal<Boolean>() {
+                    @Override public void run(Boolean value) {
+                        if (value) {
+                            MainUtil.sendMessage(player, C.SCHEMATIC_PASTE_SUCCESS);
+                        } else {
+                            MainUtil.sendMessage(player, C.SCHEMATIC_PASTE_FAILED);
+                        }
                     }
-                }
-            });
+                });
         }
         plotworld.getPlotManager().claimPlot(plotworld, this);
         return true;
@@ -1420,10 +1536,11 @@ public class Plot {
 
     /**
      * Register a plot and create it in the database<br>
-     *  - The plot will not be created if the owner is null<br>
-     *  - Any setting from before plot creation will not be saved until the server is stopped properly. i.e. Set any values/options after plot
-     *  creation.
-     * @param uuid the uuid of the plot owner
+     * - The plot will not be created if the owner is null<br>
+     * - Any setting from before plot creation will not be saved until the server is stopped properly. i.e. Set any values/options after plot
+     * creation.
+     *
+     * @param uuid   the uuid of the plot owner
      * @param notify
      * @return true if plot was created successfully
      */
@@ -1449,13 +1566,10 @@ public class Plot {
         this.getDenied().clear();
         this.settings = new PlotSettings();
         if (this.area.addPlot(this)) {
-            DBFunc.createPlotAndSettings(this, new Runnable() {
-                @Override
-                public void run() {
-                    PlotArea plotworld = Plot.this.area;
-                    if (notify && plotworld.AUTO_MERGE) {
-                        Plot.this.autoMerge(-1, Integer.MAX_VALUE, uuid, true);
-                    }
+            DBFunc.createPlotAndSettings(this, () -> {
+                PlotArea plotworld = Plot.this.area;
+                if (notify && plotworld.AUTO_MERGE) {
+                    Plot.this.autoMerge(-1, Integer.MAX_VALUE, uuid, true);
                 }
             });
             return true;
@@ -1465,7 +1579,7 @@ public class Plot {
 
     /**
      * Set components such as border, wall, floor.
-     *  (components are generator specific)
+     * (components are generator specific)
      */
     public boolean setComponent(String component, String blocks) {
         PlotBlock[] parsed = Configuration.BLOCKLIST.parseString(blocks);
@@ -1474,6 +1588,7 @@ public class Plot {
 
     /**
      * Retrieve the biome of the plot.
+     *
      * @return the name of the biome
      */
     public String getBiome() {
@@ -1483,6 +1598,7 @@ public class Plot {
 
     /**
      * Return the top location for the plot.
+     *
      * @return
      */
     public Location getTopAbs() {
@@ -1493,6 +1609,7 @@ public class Plot {
 
     /**
      * Return the bottom location for the plot.
+     *
      * @return
      */
     public Location getBottomAbs() {
@@ -1503,7 +1620,8 @@ public class Plot {
 
     /**
      * Swap the settings for two plots.
-     * @param plot the plot to swap data with
+     *
+     * @param plot     the plot to swap data with
      * @param whenDone the task to run at the end of this method.
      * @return
      */
@@ -1539,7 +1657,8 @@ public class Plot {
 
     /**
      * Move the settings for a plot.
-     * @param plot the plot to move
+     *
+     * @param plot     the plot to move
      * @param whenDone
      * @return
      */
@@ -1586,8 +1705,8 @@ public class Plot {
 
     /**
      * Gets the bottom location for a plot.<br>
-     *  - Does not respect mega plots<br>
-     *  - Merged plots, only the road will be considered part of the plot<br>
+     * - Does not respect mega plots<br>
+     * - Merged plots, only the road will be considered part of the plot<br>
      *
      * @return Location bottom of mega plot
      */
@@ -1607,22 +1726,22 @@ public class Plot {
 
     /**
      * Returns the top and bottom location.<br>
-     *  - If the plot is not connected, it will return its own corners<br>
-     *  - the returned locations will not necessarily correspond to claimed plots if the connected plots do not form a rectangular shape
-     * @deprecated as merged plots no longer need to be rectangular
+     * - If the plot is not connected, it will return its own corners<br>
+     * - the returned locations will not necessarily correspond to claimed plots if the connected plots do not form a rectangular shape
+     *
      * @return new Location[] { bottom, top }
+     * @deprecated as merged plots no longer need to be rectangular
      */
-    @Deprecated
-    public Location[] getCorners() {
+    @Deprecated public Location[] getCorners() {
         if (!this.isMerged()) {
-            return new Location[]{this.getBottomAbs(), this.getTopAbs()};
+            return new Location[] {this.getBottomAbs(), this.getTopAbs()};
         }
         return MainUtil.getCorners(this.getWorldName(), this.getRegions());
     }
 
     /**
      * Remove the east road section of a plot<br>
-     *  - Used when a plot is merged<br>
+     * - Used when a plot is merged<br>
      */
     public void removeRoadEast() {
         if (this.area.TYPE != 0 && this.area.TERRAIN > 1) {
@@ -1642,15 +1761,15 @@ public class Plot {
 
     /**
      * Returns the top and bottom plot id.<br>
-     *  - If the plot is not connected, it will return itself for the top/bottom<br>
-     *  - the returned ids will not necessarily correspond to claimed plots if the connected plots do not form a rectangular shape
-     * @deprecated as merged plots no longer need to be rectangular
+     * - If the plot is not connected, it will return itself for the top/bottom<br>
+     * - the returned ids will not necessarily correspond to claimed plots if the connected plots do not form a rectangular shape
+     *
      * @return new Plot[] { bottom, top }
+     * @deprecated as merged plots no longer need to be rectangular
      */
-    @Deprecated
-    public PlotId[] getCornerIds() {
+    @Deprecated public PlotId[] getCornerIds() {
         if (!this.isMerged()) {
-            return new PlotId[]{this.getId(), this.getId()};
+            return new PlotId[] {this.getId(), this.getId()};
         }
         PlotId min = new PlotId(this.getId().x, this.getId().y);
         PlotId max = new PlotId(this.getId().x, this.getId().y);
@@ -1666,36 +1785,34 @@ public class Plot {
                 max.y = current.getId().y;
             }
         }
-        return new PlotId[]{min, max};
+        return new PlotId[] {min, max};
     }
 
     /**
      * @deprecated in favor of getCorners()[0];<br>
-     * @return
      */
-    @Deprecated
-    public Location getBottom() {
+    @Deprecated public Location getBottom() {
         return this.getCorners()[0];
     }
 
     /**
-     * @deprecated in favor of getCorners()[1];
      * @return the top corner of the plot
+     * @deprecated in favor of getCorners()[1];
      */
-    @Deprecated
-    public Location getTop() {
+    @Deprecated public Location getTop() {
         return this.getCorners()[1];
     }
 
     /**
      * Swap the plot contents and settings with another location<br>
-     *  - The destination must correspond to a valid plot of equal dimensions
+     * - The destination must correspond to a valid plot of equal dimensions
+     *
+     * @param destination The other plot to swap with
+     * @param whenDone    A task to run when finished, or null
+     * @return boolean if swap was successful
      * @see ChunkManager#swap(Location, Location, Location, Location, Runnable) to swap terrain
      * @see this#swapData(Plot, Runnable) to swap plot settings
-     * @param destination The other plot to swap with
-     * @param whenDone A task to run when finished, or null
      * @see this#swapData(Plot, Runnable)
-     * @return boolean if swap was successful
      */
     public boolean swap(Plot destination, Runnable whenDone) {
         return this.move(destination, whenDone, true);
@@ -1703,9 +1820,10 @@ public class Plot {
 
     /**
      * Move the plot to an empty location<br>
-     *  - The location must be empty
+     * - The location must be empty
+     *
      * @param destination Where to move the plot
-     * @param whenDone A task to run when done, or null
+     * @param whenDone    A task to run when done, or null
      * @return if the move was successful
      */
     public boolean move(Plot destination, Runnable whenDone) {
@@ -1717,8 +1835,7 @@ public class Plot {
      *
      * @return alias if set, else id
      */
-    @Override
-    public String toString() {
+    @Override public String toString() {
         if (this.settings != null && this.settings.getAlias().length() > 1) {
             return this.settings.getAlias();
         }
@@ -1729,7 +1846,6 @@ public class Plot {
     /**
      * Remove a denied player (use DBFunc as well)<br>
      * Using the * uuid will remove all users
-     * @param uuid
      */
     public boolean removeDenied(UUID uuid) {
         if (uuid == DBFunc.everyone && !denied.contains(uuid)) {
@@ -1756,7 +1872,6 @@ public class Plot {
     /**
      * Remove a helper (use DBFunc as well)<br>
      * Using the * uuid will remove all users
-     * @param uuid
      */
     public boolean removeTrusted(UUID uuid) {
         if (uuid == DBFunc.everyone && !trusted.contains(uuid)) {
@@ -1783,7 +1898,6 @@ public class Plot {
     /**
      * Remove a trusted user (use DBFunc as well)<br>
      * Using the * uuid will remove all users
-     * @param uuid
      */
     public boolean removeMember(UUID uuid) {
         if (this.members == null) {
@@ -1812,12 +1926,12 @@ public class Plot {
 
     /**
      * Export the plot as a schematic to the configured output directory.
+     *
      * @return
      */
     public void export(final RunnableVal<Boolean> whenDone) {
         SchematicHandler.manager.getCompoundTag(this, new RunnableVal<CompoundTag>() {
-            @Override
-            public void run(final CompoundTag value) {
+            @Override public void run(final CompoundTag value) {
                 if (value == null) {
                     if (whenDone != null) {
                         whenDone.value = false;
@@ -1825,11 +1939,11 @@ public class Plot {
                     }
                 } else {
                     TaskManager.runTaskAsync(new Runnable() {
-                        @Override
-                        public void run() {
-                            String name = Plot.this.id + "," + Plot.this.area + ',' + MainUtil.getName(Plot.this.owner);
-                            boolean result =
-                                    SchematicHandler.manager.save(value, Settings.Paths.SCHEMATICS + File.separator + name + ".schematic");
+                        @Override public void run() {
+                            String name = Plot.this.id + "," + Plot.this.area + ',' + MainUtil
+                                .getName(Plot.this.owner);
+                            boolean result = SchematicHandler.manager.save(value,
+                                Settings.Paths.SCHEMATICS + File.separator + name + ".schematic");
                             if (whenDone != null) {
                                 whenDone.value = result;
                                 TaskManager.runTask(whenDone);
@@ -1843,9 +1957,10 @@ public class Plot {
 
     /**
      * Export the plot as a BO3 object<br>
-     *  - bedrock, floor and main block are ignored in their respective sections
-     *  - air is ignored
-     *  - The center is considered to be on top of the plot in the center
+     * - bedrock, floor and main block are ignored in their respective sections
+     * - air is ignored
+     * - The center is considered to be on top of the plot in the center
+     *
      * @param whenDone value will be false if exporting fails
      */
     public void exportBO3(RunnableVal<Boolean> whenDone) {
@@ -1858,12 +1973,12 @@ public class Plot {
 
     /**
      * Upload the plot as a schematic to the configured web interface.
+     *
      * @param whenDone value will be null if uploading fails
      */
     public void upload(final RunnableVal<URL> whenDone) {
         SchematicHandler.manager.getCompoundTag(this, new RunnableVal<CompoundTag>() {
-            @Override
-            public void run(CompoundTag value) {
+            @Override public void run(CompoundTag value) {
                 SchematicHandler.manager.upload(value, null, null, whenDone);
             }
         });
@@ -1871,10 +1986,10 @@ public class Plot {
 
     /**
      * Upload this plot as a world file<br>
-     *  - The mca files are each 512x512, so depending on the plot size it may also download adjacent plots<br>
-     *  - Works best when (plot width + road width) % 512 == 0<br>
-     *  @see WorldUtil
-     * @param whenDone
+     * - The mca files are each 512x512, so depending on the plot size it may also download adjacent plots<br>
+     * - Works best when (plot width + road width) % 512 == 0<br>
+     *
+     * @see WorldUtil
      */
     public void uploadWorld(RunnableVal<URL> whenDone) {
         WorldUtil.IMP.upload(this, null, null, whenDone);
@@ -1882,17 +1997,16 @@ public class Plot {
 
     /**
      * Upload this plot as a BO3<br>
-     *     - May not work on non default generator<br>
-     *     - BO3 includes flags/ignores plot main/floor block<br>
+     * - May not work on non default generator<br>
+     * - BO3 includes flags/ignores plot main/floor block<br>
+     *
      * @see BO3Handler
-     * @param whenDone
      */
     public void uploadBO3(RunnableVal<URL> whenDone) {
         BO3Handler.upload(this, null, null, whenDone);
     }
 
-    @Override
-    public boolean equals(Object obj) {
+    @Override public boolean equals(Object obj) {
         if (this == obj) {
             return true;
         }
@@ -1903,24 +2017,26 @@ public class Plot {
             return false;
         }
         Plot other = (Plot) obj;
-        return this.hashCode() == other.hashCode() && this.id.equals(other.id) && this.area == other.area;
+        return this.hashCode() == other.hashCode() && this.id.equals(other.id)
+            && this.area == other.area;
     }
 
     /**
      * Get the plot hashcode<br>
      * Note: The hashcode is unique if:<br>
-     *     - Plots are in the same world<br>
-     *     - The x,z coordinates are between Short.MIN_VALUE and Short.MAX_VALUE<br>
+     * - Plots are in the same world<br>
+     * - The x,z coordinates are between Short.MIN_VALUE and Short.MAX_VALUE<br>
+     *
      * @return integer.
      */
-    @Override
-    public int hashCode() {
+    @Override public int hashCode() {
         return this.id.hashCode();
     }
 
     /**
      * Get the flags specific to this plot<br>
-     *  - Does not take default flags into account<br>
+     * - Does not take default flags into account<br>
+     *
      * @return
      */
     public HashMap<Flag<?>, Object> getFlags() {
@@ -1929,7 +2045,6 @@ public class Plot {
 
     /**
      * Set a flag for this plot.
-     * @param flags
      */
     public void setFlags(HashMap<Flag<?>, Object> flags) {
         FlagManager.setPlotFlags(this, flags);
@@ -1937,7 +2052,8 @@ public class Plot {
 
     /**
      * Get the plot alias.
-     *  - Returns an empty string if no alias is set
+     * - Returns an empty string if no alias is set
+     *
      * @return The plot alias
      */
     public String getAlias() {
@@ -1949,6 +2065,7 @@ public class Plot {
 
     /**
      * Set the plot alias.
+     *
      * @param alias The alias
      */
     public void setAlias(String alias) {
@@ -1967,23 +2084,27 @@ public class Plot {
 
     /**
      * Set the raw merge data<br>
-     *  - Updates DB<br>
-     *  - Does not modify terrain<br>
+     * - Updates DB<br>
+     * - Does not modify terrain<br>
      * ----------<br>
      * 0 = north<br>
      * 1 = east<br>
      * 2 = south<br>
      * 3 = west<br>
      * ----------<br>
-     * @param direction
-     * @param value
      */
     public void setMerged(int direction, boolean value) {
         if (this.getSettings().setMerged(direction, value)) {
             if (value) {
                 Plot other = this.getRelative(direction).getBasePlot(false);
                 if (!other.equals(this.getBasePlot(false))) {
-                    Plot base = other.id.y < this.id.y || other.id.y == this.id.y && other.id.x < this.id.x ? other : this.origin;
+                    Plot base;
+                    if (other.id.y < this.id.y
+                        || other.id.y == this.id.y && other.id.x < this.id.x) {
+                        base = other;
+                    } else {
+                        base = this.origin;
+                    }
                     this.origin.origin = base;
                     other.origin = base;
                     this.origin = base;
@@ -2003,6 +2124,7 @@ public class Plot {
 
     /**
      * Get the merged array.
+     *
      * @return boolean [ north, east, south, west ]
      */
     public boolean[] getMerged() {
@@ -2011,8 +2133,8 @@ public class Plot {
 
     /**
      * Set the raw merge data<br>
-     *  - Updates DB<br>
-     *  - Does not modify terrain<br>
+     * - Updates DB<br>
+     * - Does not modify terrain<br>
      * Get if the plot is merged in a direction<br>
      * ----------<br>
      * 0 = north<br>
@@ -2021,6 +2143,7 @@ public class Plot {
      * 3 = west<br>
      * ----------<br>
      * Note: Diagonal merging (4-7) must be done by merging the corresponding plots.
+     *
      * @param merged
      */
     public void setMerged(boolean[] merged) {
@@ -2040,9 +2163,10 @@ public class Plot {
 
     /**
      * Get the set home location or 0,0,0 if no location is set<br>
-     *  - Does not take the default home location into account
-     * @see #getHome()
+     * - Does not take the default home location into account
+     *
      * @return
+     * @see #getHome()
      */
     public BlockLoc getPosition() {
         return this.getSettings().getPosition();
@@ -2050,13 +2174,15 @@ public class Plot {
 
     /**
      * Check if a plot can be claimed by the provided player.
+     *
      * @param player the claiming player
      * @return
      */
     public boolean canClaim(PlotPlayer player) {
         PlotCluster cluster = this.getCluster();
         if (cluster != null && player != null) {
-            if (!cluster.isAdded(player.getUUID()) && !Permissions.hasPermission(player, "plots.admin.command.claim")) {
+            if (!cluster.isAdded(player.getUUID()) && !Permissions
+                .hasPermission(player, "plots.admin.command.claim")) {
                 return false;
             }
         }
@@ -2066,6 +2192,7 @@ public class Plot {
     /**
      * Guess the owner of a plot either by the value in memory, or the sign data<br>
      * Note: Recovering from sign information is useful if e.g. PlotMe conversion wasn't successful
+     *
      * @return UUID
      */
     public UUID guessOwner() {
@@ -2078,8 +2205,7 @@ public class Plot {
         try {
             final Location loc = this.getManager().getSignLoc(this.area, this);
             String[] lines = TaskManager.IMP.sync(new RunnableVal<String[]>() {
-                @Override
-                public void run(String[] value) {
+                @Override public void run(String[] value) {
                     ChunkManager.manager.loadChunk(loc.getWorld(), loc.getChunkLoc(), false);
                     this.value = WorldUtil.IMP.getSign(loc);
                 }
@@ -2117,7 +2243,8 @@ public class Plot {
                         }
                     }
                 }
-                this.owner = UUID.nameUUIDFromBytes(("OfflinePlayer:" + name).getBytes(StandardCharsets.UTF_8));
+                this.owner = UUID.nameUUIDFromBytes(
+                    ("OfflinePlayer:" + name).getBytes(StandardCharsets.UTF_8));
                 break;
             }
             if (this.hasOwner()) {
@@ -2131,7 +2258,7 @@ public class Plot {
 
     /**
      * Remove the south road section of a plot<br>
-     *  - Used when a plot is merged<br>
+     * - Used when a plot is merged<br>
      */
     public void removeRoadSouth() {
         if (this.area.TYPE != 0 && this.area.TERRAIN > 1) {
@@ -2151,14 +2278,15 @@ public class Plot {
 
     /**
      * Auto merge a plot in a specific direction<br>
-     * @param dir The direction to merge<br>
-     * -1 = All directions<br>
-     * 0 = north<br>
-     * 1 = east<br>
-     * 2 = south<br>
-     * 3 = west<br>
-     * @param max The max number of merges to do
-     * @param uuid The UUID it is allowed to merge with
+     *
+     * @param dir         The direction to merge<br>
+     *                    -1 = All directions<br>
+     *                    0 = north<br>
+     *                    1 = east<br>
+     *                    2 = south<br>
+     *                    3 = west<br>
+     * @param max         The max number of merges to do
+     * @param uuid        The UUID it is allowed to merge with
      * @param removeRoads Whether to remove roads
      * @return true if a merge takes place
      */
@@ -2166,7 +2294,7 @@ public class Plot {
         if (this.owner == null) {
             return false;
         }
-        if(!EventUtil.manager.callMerge(this, dir, max)) {
+        if (!EventUtil.manager.callMerge(this, dir, max)) {
             return false;
         }
         HashSet<Plot> visited = new HashSet<>();
@@ -2186,9 +2314,10 @@ public class Plot {
             Set<Plot> plots;
             if ((dir == -1 || dir == 0) && !current.getMerged(0)) {
                 Plot other = current.getRelative(0);
-                if (other != null && other.isOwner(uuid)
-                        && (other.getBasePlot(false).equals(current.getBasePlot(false))
-                        || (plots = other.getConnectedPlots()).size() <= max && frontier.addAll(plots) && (max -= plots.size()) != -1)) {
+                if (other != null && other.isOwner(uuid) && (
+                    other.getBasePlot(false).equals(current.getBasePlot(false))
+                        || (plots = other.getConnectedPlots()).size() <= max && frontier
+                        .addAll(plots) && (max -= plots.size()) != -1)) {
                     current.mergePlot(other, removeRoads);
                     merged.add(current.getId());
                     merged.add(other.getId());
@@ -2197,9 +2326,10 @@ public class Plot {
             }
             if (max >= 0 && (dir == -1 || dir == 1) && !current.getMerged(1)) {
                 Plot other = current.getRelative(1);
-                if (other != null && other.isOwner(uuid)
-                        && (other.getBasePlot(false).equals(current.getBasePlot(false))
-                        || (plots = other.getConnectedPlots()).size() <= max && frontier.addAll(plots) && (max -= plots.size()) != -1)) {
+                if (other != null && other.isOwner(uuid) && (
+                    other.getBasePlot(false).equals(current.getBasePlot(false))
+                        || (plots = other.getConnectedPlots()).size() <= max && frontier
+                        .addAll(plots) && (max -= plots.size()) != -1)) {
                     current.mergePlot(other, removeRoads);
                     merged.add(current.getId());
                     merged.add(other.getId());
@@ -2208,9 +2338,10 @@ public class Plot {
             }
             if (max >= 0 && (dir == -1 || dir == 2) && !current.getMerged(2)) {
                 Plot other = current.getRelative(2);
-                if (other != null && other.isOwner(uuid)
-                        && (other.getBasePlot(false).equals(current.getBasePlot(false))
-                        || (plots = other.getConnectedPlots()).size() <= max && frontier.addAll(plots) && (max -= plots.size()) != -1)) {
+                if (other != null && other.isOwner(uuid) && (
+                    other.getBasePlot(false).equals(current.getBasePlot(false))
+                        || (plots = other.getConnectedPlots()).size() <= max && frontier
+                        .addAll(plots) && (max -= plots.size()) != -1)) {
                     current.mergePlot(other, removeRoads);
                     merged.add(current.getId());
                     merged.add(other.getId());
@@ -2219,9 +2350,10 @@ public class Plot {
             }
             if (max >= 0 && (dir == -1 || dir == 3) && !current.getMerged(3)) {
                 Plot other = current.getRelative(3);
-                if (other != null && other.isOwner(uuid)
-                        && (other.getBasePlot(false).equals(current.getBasePlot(false))
-                        || (plots = other.getConnectedPlots()).size() <= max && frontier.addAll(plots) && (max -= plots.size()) != -1)) {
+                if (other != null && other.isOwner(uuid) && (
+                    other.getBasePlot(false).equals(current.getBasePlot(false))
+                        || (plots = other.getConnectedPlots()).size() <= max && frontier
+                        .addAll(plots) && (max -= plots.size()) != -1)) {
                     current.mergePlot(other, removeRoads);
                     merged.add(current.getId());
                     merged.add(other.getId());
@@ -2238,7 +2370,8 @@ public class Plot {
 
     /**
      * Merge the plot settings<br>
-     *  - Used when a plot is merged<br>
+     * - Used when a plot is merged<br>
+     *
      * @param b
      */
     public void mergeData(Plot b) {
@@ -2251,7 +2384,12 @@ public class Plot {
             } else {
                 flags2.putAll(flags1);
             }
-            HashMap<Flag<?>, Object> net = (greater ? flags1 : flags2);
+            HashMap<Flag<?>, Object> net;
+            if (greater) {
+                net = flags1;
+            } else {
+                net = flags2;
+            }
             this.setFlags(net);
             b.setFlags(net);
         }
@@ -2303,6 +2441,7 @@ public class Plot {
     /**
      * Get the plot in a relative location<br>
      * Note: May be null if the partial plot area does not include the relative location
+     *
      * @param x
      * @param y
      * @return Plot
@@ -2322,6 +2461,7 @@ public class Plot {
      * 2 = south<br>
      * 3 = west<br>
      * Note: May be null if the partial plot area does not include the relative location
+     *
      * @param direction
      * @return
      */
@@ -2331,7 +2471,8 @@ public class Plot {
 
     /**
      * Get a set of plots connected (and including) this plot<br>
-     *  - This result is cached globally
+     * - This result is cached globally
+     *
      * @return
      */
     public Set<Plot> getConnectedPlots() {
@@ -2463,20 +2604,22 @@ public class Plot {
 
     /**
      * This will combine each plot into effective rectangular regions<br>
-     *  - This result is cached globally<br>
-     *  - Useful for handling non rectangular shapes
+     * - This result is cached globally<br>
+     * - Useful for handling non rectangular shapes
+     *
      * @return
      */
-    public HashSet<RegionWrapper> getRegions() {
+    @Nonnull public HashSet<RegionWrapper> getRegions() {
         if (regions_cache != null && connected_cache != null && connected_cache.contains(this)) {
             return regions_cache;
         }
         if (!this.isMerged()) {
             Location pos1 = this.getBottomAbs();
             Location pos2 = this.getTopAbs();
-            connected_cache = new HashSet<>(Collections.singletonList(this));
-            regions_cache = new HashSet<>(1);
-            regions_cache.add(new RegionWrapper(pos1.getX(), pos2.getX(), pos1.getY(), pos2.getY(), pos1.getZ(), pos2.getZ()));
+            connected_cache = Sets.newHashSet(this);
+            regions_cache = Sets.newHashSet(
+                new RegionWrapper(pos1.getX(), pos2.getX(), pos1.getY(), pos2.getY(), pos1.getZ(),
+                    pos2.getZ()));
             return regions_cache;
         }
         Set<Plot> plots = this.getConnectedPlots();
@@ -2491,7 +2634,8 @@ public class Plot {
             PlotId top = new PlotId(current.getId().x, current.getId().y);
             while (merge) {
                 merge = false;
-                ArrayList<PlotId> ids = MainUtil.getPlotSelectionIds(new PlotId(bot.x, bot.y - 1), new PlotId(top.x, bot.y - 1));
+                ArrayList<PlotId> ids = MainUtil.getPlotSelectionIds(new PlotId(bot.x, bot.y - 1),
+                    new PlotId(top.x, bot.y - 1));
                 boolean tmp = true;
                 for (PlotId id : ids) {
                     Plot plot = this.area.getPlotAbs(id);
@@ -2503,7 +2647,8 @@ public class Plot {
                     merge = true;
                     bot.y--;
                 }
-                ids = MainUtil.getPlotSelectionIds(new PlotId(top.x + 1, bot.y), new PlotId(top.x + 1, top.y));
+                ids = MainUtil.getPlotSelectionIds(new PlotId(top.x + 1, bot.y),
+                    new PlotId(top.x + 1, top.y));
                 tmp = true;
                 for (PlotId id : ids) {
                     Plot plot = this.area.getPlotAbs(id);
@@ -2515,7 +2660,8 @@ public class Plot {
                     merge = true;
                     top.x++;
                 }
-                ids = MainUtil.getPlotSelectionIds(new PlotId(bot.x, top.y + 1), new PlotId(top.x, top.y + 1));
+                ids = MainUtil.getPlotSelectionIds(new PlotId(bot.x, top.y + 1),
+                    new PlotId(top.x, top.y + 1));
                 tmp = true;
                 for (PlotId id : ids) {
                     Plot plot = this.area.getPlotAbs(id);
@@ -2527,7 +2673,8 @@ public class Plot {
                     merge = true;
                     top.y++;
                 }
-                ids = MainUtil.getPlotSelectionIds(new PlotId(bot.x - 1, bot.y), new PlotId(bot.x - 1, top.y));
+                ids = MainUtil.getPlotSelectionIds(new PlotId(bot.x - 1, bot.y),
+                    new PlotId(bot.x - 1, top.y));
                 tmp = true;
                 for (PlotId id : ids) {
                     Plot plot = this.area.getPlotAbs(id);
@@ -2542,9 +2689,7 @@ public class Plot {
             }
             Location gtopabs = this.area.getPlotAbs(top).getTopAbs();
             Location gbotabs = this.area.getPlotAbs(bot).getBottomAbs();
-            for (PlotId id : MainUtil.getPlotSelectionIds(bot, top)) {
-                visited.add(id);
-            }
+            visited.addAll(MainUtil.getPlotSelectionIds(bot, top));
             for (int x = bot.x; x <= top.x; x++) {
                 Plot plot = this.area.getPlotAbs(new PlotId(x, top.y));
                 if (plot.getMerged(2)) {
@@ -2552,9 +2697,12 @@ public class Plot {
                     Location toploc = plot.getExtendedTopAbs();
                     Location botabs = plot.getBottomAbs();
                     Location topabs = plot.getTopAbs();
-                    regions.add(new RegionWrapper(botabs.getX(), topabs.getX(), topabs.getZ() + 1, toploc.getZ()));
+                    regions.add(new RegionWrapper(botabs.getX(), topabs.getX(), topabs.getZ() + 1,
+                        toploc.getZ()));
                     if (plot.getMerged(5)) {
-                        regions.add(new RegionWrapper(topabs.getX() + 1, toploc.getX(), topabs.getZ() + 1, toploc.getZ()));
+                        regions.add(
+                            new RegionWrapper(topabs.getX() + 1, toploc.getX(), topabs.getZ() + 1,
+                                toploc.getZ()));
                         // intersection
                     }
                 }
@@ -2567,20 +2715,25 @@ public class Plot {
                     Location toploc = plot.getExtendedTopAbs();
                     Location botabs = plot.getBottomAbs();
                     Location topabs = plot.getTopAbs();
-                    regions.add(new RegionWrapper(topabs.getX() + 1, toploc.getX(), botabs.getZ(), topabs.getZ()));
+                    regions.add(new RegionWrapper(topabs.getX() + 1, toploc.getX(), botabs.getZ(),
+                        topabs.getZ()));
                     if (plot.getMerged(5)) {
-                        regions.add(new RegionWrapper(topabs.getX() + 1, toploc.getX(), topabs.getZ() + 1, toploc.getZ()));
+                        regions.add(
+                            new RegionWrapper(topabs.getX() + 1, toploc.getX(), topabs.getZ() + 1,
+                                toploc.getZ()));
                         // intersection
                     }
                 }
             }
-            regions.add(new RegionWrapper(gbotabs.getX(), gtopabs.getX(), gbotabs.getZ(), gtopabs.getZ()));
+            regions.add(
+                new RegionWrapper(gbotabs.getX(), gtopabs.getX(), gbotabs.getZ(), gtopabs.getZ()));
         }
         return regions;
     }
 
     /**
      * Attempt to find the largest rectangular region in a plot (as plots can form non rectangular shapes)
+     *
      * @return
      */
     public RegionWrapper getLargestRegion() {
@@ -2588,7 +2741,8 @@ public class Plot {
         RegionWrapper max = null;
         double area = Double.NEGATIVE_INFINITY;
         for (RegionWrapper region : regions) {
-            double current = (region.maxX - (double) region.minX + 1) * (region.maxZ - (double) region.minZ + 1);
+            double current =
+                (region.maxX - (double) region.minX + 1) * (region.maxZ - (double) region.minZ + 1);
             if (current > area) {
                 max = region;
                 area = current;
@@ -2599,12 +2753,11 @@ public class Plot {
 
     /**
      * Do the plot entry tasks for each player in the plot<br>
-     *  - Usually called when the plot state changes (unclaimed/claimed/flag change etc)
+     * - Usually called when the plot state changes (unclaimed/claimed/flag change etc)
      */
     public void reEnter() {
         TaskManager.runTaskLater(new Runnable() {
-            @Override
-            public void run() {
+            @Override public void run() {
                 for (PlotPlayer pp : Plot.this.getPlayersInPlot()) {
                     PlotListener.plotExit(pp, Plot.this);
                     PlotListener.plotEntry(pp, Plot.this);
@@ -2615,12 +2768,14 @@ public class Plot {
 
     /**
      * Get all the corners of the plot (supports non-rectangular shapes).
+     *
      * @return A list of the plot corners
      */
     public List<Location> getAllCorners() {
         Area area = new Area();
         for (RegionWrapper region : this.getRegions()) {
-            Rectangle2D rect = new Rectangle2D.Double(region.minX - 0.6, region.minZ - 0.6, region.maxX - region.minX + 1.2, region.maxZ - region.minZ + 1.2);
+            Rectangle2D rect = new Rectangle2D.Double(region.minX - 0.6, region.minZ - 0.6,
+                region.maxX - region.minX + 1.2, region.maxZ - region.minZ + 1.2);
             Area rectArea = new Area(rect);
             area.add(rectArea);
         }
@@ -2639,6 +2794,7 @@ public class Plot {
 
     /**
      * Teleport a player to a plot and send them the teleport message.
+     *
      * @param player the player
      * @return if the teleport succeeded
      */
@@ -2652,7 +2808,8 @@ public class Plot {
             } else {
                 location = this.getDefaultHome(false);
             }
-            if (Settings.Teleport.DELAY == 0 || Permissions.hasPermission(player, "plots.teleport.delay.bypass")) {
+            if (Settings.Teleport.DELAY == 0 || Permissions
+                .hasPermission(player, "plots.teleport.delay.bypass")) {
                 MainUtil.sendMessage(player, C.TELEPORTED_TO_PLOT);
                 player.teleport(location);
                 return true;
@@ -2661,8 +2818,7 @@ public class Plot {
             final String name = player.getName();
             TaskManager.TELEPORT_QUEUE.add(name);
             TaskManager.runTaskLater(new Runnable() {
-                @Override
-                public void run() {
+                @Override public void run() {
                     if (!TaskManager.TELEPORT_QUEUE.contains(name)) {
                         MainUtil.sendMessage(player, C.TELEPORT_FAILED);
                         return;
@@ -2696,14 +2852,16 @@ public class Plot {
 
     /**
      * Set a component for a plot to the provided blocks<br>
-     *  - E.g. floor, wall, border etc.<br>
-     *  - The available components depend on the generator being used<br>
+     * - E.g. floor, wall, border etc.<br>
+     * - The available components depend on the generator being used<br>
+     *
      * @param component
      * @param blocks
      * @return
      */
     public boolean setComponent(String component, PlotBlock[] blocks) {
-        if (StringMan.isEqualToAny(component, getManager().getPlotComponents(this.area, this.getId()))) {
+        if (StringMan
+            .isEqualToAny(component, getManager().getPlotComponents(this.area, this.getId()))) {
             EventUtil.manager.callComponentSet(this, component);
         }
         return this.getManager().setComponent(this.area, this.getId(), component, blocks);
@@ -2712,7 +2870,8 @@ public class Plot {
     public int getDistanceFromOrigin() {
         Location bot = getManager().getPlotBottomLocAbs(this.area, id);
         Location top = getManager().getPlotTopLocAbs(this.area, id);
-        return Math.max(Math.max(Math.abs(bot.getX()), Math.abs(bot.getZ())), Math.max(Math.abs(top.getX()), Math.abs(top.getZ())));
+        return Math.max(Math.max(Math.abs(bot.getX()), Math.abs(bot.getZ())),
+            Math.max(Math.abs(top.getX()), Math.abs(top.getZ())));
     }
 
     /**
@@ -2793,13 +2952,15 @@ public class Plot {
 
     /**
      * Move a plot physically, as well as the corresponding settings.
+     *
      * @param destination
      * @param whenDone
      * @param allowSwap
      * @return
      */
     public boolean move(final Plot destination, final Runnable whenDone, boolean allowSwap) {
-        final PlotId offset = new PlotId(destination.getId().x - this.getId().x, destination.getId().y - this.getId().y);
+        final PlotId offset = new PlotId(destination.getId().x - this.getId().x,
+            destination.getId().y - this.getId().y);
         Location db = destination.getBottomAbs();
         Location ob = this.getBottomAbs();
         final int offsetX = db.getX() - ob.getX();
@@ -2831,57 +2992,52 @@ public class Plot {
             Plot other = plot.getRelative(destination.getArea(), offset.x, offset.y);
             plot.swapData(other, null);
         }
-        // copy terrain
-        Runnable move = new Runnable() {
-            @Override
-            public void run() {
-                if (regions.isEmpty()) {
-                    Plot plot = destination.getRelative(0, 0);
-                    for (Plot current : plot.getConnectedPlots()) {
-                        getManager().claimPlot(current.getArea(), current);
-                        Plot originPlot = originArea.getPlotAbs(new PlotId(current.id.x - offset.x, current.id.y - offset.y));
-                        originPlot.getManager().unclaimPlot(originArea, originPlot, null);
-                    }
-                    plot.setSign();
-                    TaskManager.runTask(whenDone);
-                    return;
-                }
-                final Runnable task = this;
-                RegionWrapper region = regions.poll();
-                Location[] corners = region.getCorners(getWorldName());
-                final Location pos1 = corners[0];
-                final Location pos2 = corners[1];
-                Location newPos = pos1.clone().add(offsetX, 0, offsetZ);
-                newPos.setWorld(destination.getWorldName());
-                ChunkManager.manager.copyRegion(pos1, pos2, newPos, new Runnable() {
-                    @Override
-                    public void run() {
-                        ChunkManager.manager.regenerateRegion(pos1, pos2, false, task);
-                    }
-                });
-            }
-        };
-        Runnable swap = new Runnable() {
-            @Override
-            public void run() {
-                if (regions.isEmpty()) {
-                    TaskManager.runTask(whenDone);
-                    return;
-                }
-                RegionWrapper region = regions.poll();
-                Location[] corners = region.getCorners(getWorldName());
-                Location pos1 = corners[0];
-                Location pos2 = corners[1];
-                Location pos3 = pos1.clone().add(offsetX, 0, offsetZ);
-                Location pos4 = pos2.clone().add(offsetX, 0, offsetZ);
-                pos3.setWorld(destination.getWorldName());
-                pos4.setWorld(destination.getWorldName());
-                ChunkManager.manager.swap(pos1, pos2, pos3, pos4, this);
-            }
-        };
         if (occupied) {
+            Runnable swap = new Runnable() {
+                @Override public void run() {
+                    if (regions.isEmpty()) {
+                        TaskManager.runTask(whenDone);
+                        return;
+                    }
+                    RegionWrapper region = regions.poll();
+                    Location[] corners = region.getCorners(getWorldName());
+                    Location pos1 = corners[0];
+                    Location pos2 = corners[1];
+                    Location pos3 = pos1.clone().add(offsetX, 0, offsetZ);
+                    Location pos4 = pos2.clone().add(offsetX, 0, offsetZ);
+                    pos3.setWorld(destination.getWorldName());
+                    pos4.setWorld(destination.getWorldName());
+                    ChunkManager.manager.swap(pos1, pos2, pos3, pos4, this);
+                }
+            };
             swap.run();
         } else {
+            //copy terrain
+            Runnable move = new Runnable() {
+                @Override public void run() {
+                    if (regions.isEmpty()) {
+                        Plot plot = destination.getRelative(0, 0);
+                        for (Plot current : plot.getConnectedPlots()) {
+                            getManager().claimPlot(current.getArea(), current);
+                            Plot originPlot = originArea.getPlotAbs(
+                                new PlotId(current.id.x - offset.x, current.id.y - offset.y));
+                            originPlot.getManager().unclaimPlot(originArea, originPlot, null);
+                        }
+                        plot.setSign();
+                        TaskManager.runTask(whenDone);
+                        return;
+                    }
+                    final Runnable task = this;
+                    RegionWrapper region = regions.poll();
+                    Location[] corners = region.getCorners(getWorldName());
+                    final Location pos1 = corners[0];
+                    final Location pos2 = corners[1];
+                    Location newPos = pos1.clone().add(offsetX, 0, offsetZ);
+                    newPos.setWorld(destination.getWorldName());
+                    ChunkManager.manager.copyRegion(pos1, pos2, newPos,
+                        () -> ChunkManager.manager.regenerateRegion(pos1, pos2, false, task));
+                }
+            };
             move.run();
         }
         return true;
@@ -2889,12 +3045,14 @@ public class Plot {
 
     /**
      * Copy a plot to a location, both physically and the settings
+     *
      * @param destination
      * @param whenDone
      * @return
      */
     public boolean copy(final Plot destination, final Runnable whenDone) {
-        PlotId offset = new PlotId(destination.getId().x - this.getId().x, destination.getId().y - this.getId().y);
+        PlotId offset = new PlotId(destination.getId().x - this.getId().x,
+            destination.getId().y - this.getId().y);
         Location db = destination.getBottomAbs();
         Location ob = this.getBottomAbs();
         final int offsetX = db.getX() - ob.getX();
@@ -2946,8 +3104,7 @@ public class Plot {
         // copy terrain
         final ArrayDeque<RegionWrapper> regions = new ArrayDeque<>(this.getRegions());
         Runnable run = new Runnable() {
-            @Override
-            public void run() {
+            @Override public void run() {
                 if (regions.isEmpty()) {
                     for (Plot current : getConnectedPlots()) {
                         destination.getManager().claimPlot(destination.getArea(), destination);
