@@ -15,6 +15,7 @@ import com.intellectualcrafters.plot.object.PlotCluster;
 import com.intellectualcrafters.plot.object.PlotId;
 import com.intellectualcrafters.plot.object.PlotPlayer;
 import com.intellectualcrafters.plot.object.Rating;
+import com.intellectualcrafters.plot.object.worlds.SinglePlotArea;
 import com.intellectualcrafters.plot.util.expiry.ExpireManager;
 import com.plotsquared.listener.PlayerBlockEventType;
 
@@ -23,19 +24,21 @@ import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.UUID;
 
-public abstract class EventUtil {
+import static com.plotsquared.listener.PlayerBlockEventType.*;
+
+@SuppressWarnings("Guava") public abstract class EventUtil {
 
     public static EventUtil manager = null;
 
     /**
      * Submit a plot rate event and return the updated rating
+     *
      * @param player Player that rated the plot
-     * @param plot Plot that was rated
+     * @param plot   Plot that was rated
      * @param rating Rating given to the plot
      * @return Updated rating or null if the event was cancelled
      */
-    @Nullable
-    public abstract Rating callRating(PlotPlayer player, Plot plot, Rating rating);
+    @Nullable public abstract Rating callRating(PlotPlayer player, Plot plot, Rating rating);
 
     public abstract boolean callClaim(PlotPlayer player, Plot plot, boolean auto);
 
@@ -69,14 +72,15 @@ public abstract class EventUtil {
 
     public abstract void callMember(PlotPlayer initiator, Plot plot, UUID player, boolean added);
 
-    public abstract boolean callOwnerChange(PlotPlayer initiator, Plot plot, UUID newOwner, UUID oldOwner, boolean hasOldOwner);
+    public abstract boolean callOwnerChange(PlotPlayer initiator, Plot plot, UUID newOwner,
+        UUID oldOwner, boolean hasOldOwner);
 
     public void doJoinTask(final PlotPlayer player) {
         if (player == null) {
             return; //possible future warning message to figure out where we are retrieving null
         }
         if (ExpireManager.IMP != null) {
-                ExpireManager.IMP.handleJoin(player);
+            ExpireManager.IMP.handleJoin(player);
         }
         if (PS.get().worldedit != null) {
             if (player.getAttribute("worldedit")) {
@@ -84,120 +88,97 @@ public abstract class EventUtil {
             }
         }
         final Plot plot = player.getCurrentPlot();
-        if (Settings.Teleport.ON_LOGIN && plot != null) {
-            TaskManager.runTask(new Runnable() {
-                @Override
-                public void run() {
-                    plot.teleportPlayer(player);
-                }
-            });
-            MainUtil.sendMessage(player, C.TELEPORTED_TO_ROAD.f() + " (on-login) " + "(" + plot.getId().x + ";" + plot.getId().y + ")");
+        if (Settings.Teleport.ON_LOGIN && plot != null && !(plot.getArea() instanceof SinglePlotArea)) {
+            TaskManager.runTask(() -> plot.teleportPlayer(player));
+            MainUtil.sendMessage(player,
+                C.TELEPORTED_TO_ROAD.f() + " (on-login) " + "(" + plot.getId().x + ";" + plot
+                    .getId().y + ")");
         }
     }
 
     public void doRespawnTask(final PlotPlayer player) {
         final Plot plot = player.getCurrentPlot();
         if (Settings.Teleport.ON_DEATH && plot != null) {
-            TaskManager.runTask(new Runnable() {
-                @Override
-                public void run() {
-                    plot.teleportPlayer(player);
-                }
-            });
+            TaskManager.runTask(() -> plot.teleportPlayer(player));
             MainUtil.sendMessage(player, C.TELEPORTED_TO_ROAD);
         }
     }
 
-    public boolean checkPlayerBlockEvent(PlotPlayer player, PlayerBlockEventType type, Location location, LazyBlock block, boolean notifyPerms) {
-        PlotArea area = PS.get().getPlotAreaAbs(location);
-        Plot plot;
-        if (area != null) {
-            plot = area.getPlot(location);
-        } else {
-            plot = null;
-        }
-        if (plot == null) {
-            if (area == null) {
-                return true;
-            }
-        } else if (plot.isAdded(player.getUUID())) {
+    public boolean checkPlayerBlockEvent(PlotPlayer player, PlayerBlockEventType type,
+        Location location, LazyBlock block, boolean notifyPerms) {
+        if (type == EAT || type == READ) {
             return true;
         }
-        switch (type) {
-            case TELEPORT_OBJECT:
-                return false;
-            case EAT:
-            case READ:
+        if (type == TELEPORT_OBJECT) {
+            return false;
+        }
+        PlotArea area = player.getPlotAreaAbs();
+        assert area != null;
+        Plot plot = location.getPlot();
+        if (plot != null) {
+            if (plot.isAdded(player.getUUID())) {
                 return true;
+            }
+            if (!plot.hasOwner()) {
+                return Permissions
+                    .hasPermission(player, C.PERMISSION_ADMIN_INTERACT_UNOWNED.s(), notifyPerms);
+            }
+        } else {
+            return Permissions
+                .hasPermission(player, C.PERMISSION_ADMIN_INTERACT_ROAD.s(), notifyPerms);
+
+        }
+        switch (type) {
             case BREAK_BLOCK:
-                if (plot == null) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_ROAD.s(), notifyPerms);
-                }
-                if (!plot.hasOwner()) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_UNOWNED.s(), notifyPerms);
-                }
                 Optional<HashSet<PlotBlock>> use = plot.getFlag(Flags.USE);
                 if (use.isPresent()) {
                     HashSet<PlotBlock> value = use.get();
-                    if (value.contains(PlotBlock.EVERYTHING) || value.contains(block.getPlotBlock())) {
+                    if (value.contains(PlotBlock.EVERYTHING) || value
+                        .contains(block.getPlotBlock())) {
                         return true;
                     }
                 }
                 Optional<HashSet<PlotBlock>> destroy = plot.getFlag(Flags.BREAK);
                 if (destroy.isPresent()) {
                     HashSet<PlotBlock> value = destroy.get();
-                    if (value.contains(PlotBlock.EVERYTHING) || value.contains(block.getPlotBlock())) {
+                    if (value.contains(PlotBlock.EVERYTHING) || value
+                        .contains(block.getPlotBlock())) {
                         return true;
                     }
                 }
-                if (Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false)) {
+                if (Permissions
+                    .hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false)) {
                     return true;
                 }
-                return !(!notifyPerms || MainUtil.sendMessage(player, C.FLAG_TUTORIAL_USAGE, C.FLAG_USE.s() + '/' + C.FLAG_BREAK.s()));
+                return !(!notifyPerms || MainUtil.sendMessage(player, C.FLAG_TUTORIAL_USAGE,
+                    C.FLAG_USE.s() + '/' + C.FLAG_BREAK.s()));
             case BREAK_HANGING:
-                if (plot == null) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_ROAD.s(), notifyPerms);
-                }
                 if (plot.getFlag(Flags.HANGING_BREAK).or(false)) {
                     return true;
                 }
-                if (plot.hasOwner()) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false) || !(!notifyPerms || MainUtil.sendMessage(player, C.FLAG_TUTORIAL_USAGE, C.FLAG_HANGING_BREAK.s()));
-                }
-                return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_UNOWNED.s(), notifyPerms);
+                return
+                    Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false)
+                        || !(!notifyPerms || MainUtil
+                        .sendMessage(player, C.FLAG_TUTORIAL_USAGE, C.FLAG_HANGING_BREAK.s()));
             case BREAK_MISC:
-                if (plot == null) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_ROAD.s(), notifyPerms);
-                }
                 if (plot.getFlag(Flags.MISC_BREAK).or(false)) {
                     return true;
                 }
-                if (plot.hasOwner()) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false) || !(!notifyPerms || MainUtil
-                            .sendMessage(player, C.FLAG_TUTORIAL_USAGE, C.FLAG_MISC_BREAK.s()));
-                }
-                return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_UNOWNED.s(), notifyPerms);
+                return
+                    Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false)
+                        || !(!notifyPerms || MainUtil
+                        .sendMessage(player, C.FLAG_TUTORIAL_USAGE, C.FLAG_MISC_BREAK.s()));
             case BREAK_VEHICLE:
-                if (plot == null) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_ROAD.s(), notifyPerms);
-                }
                 if (plot.getFlag(Flags.VEHICLE_BREAK).or(false)) {
                     return true;
                 }
-                if (plot.hasOwner()) {
-                    if (Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false)) {
-                        return true;
-                    }
-                    return !(!notifyPerms || MainUtil.sendMessage(player, C.FLAG_TUTORIAL_USAGE, C.FLAG_VEHICLE_BREAK.s()));
+                if (Permissions
+                    .hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false)) {
+                    return true;
                 }
-                return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_UNOWNED.s(), notifyPerms);
+                return !(!notifyPerms || MainUtil
+                    .sendMessage(player, C.FLAG_TUTORIAL_USAGE, C.FLAG_VEHICLE_BREAK.s()));
             case INTERACT_BLOCK: {
-                if (plot == null) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_ROAD.s(), notifyPerms);
-                }
-                if (!plot.hasOwner()) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_UNOWNED.s(), notifyPerms);
-                }
                 Optional<HashSet<PlotBlock>> flagValue = plot.getFlag(Flags.USE);
                 HashSet<PlotBlock> value;
                 if (flagValue.isPresent()) {
@@ -205,18 +186,19 @@ public abstract class EventUtil {
                 } else {
                     value = null;
                 }
-                if (value == null || !value.contains(PlotBlock.EVERYTHING) && !value.contains(block.getPlotBlock())) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false) || !(!notifyPerms || MainUtil
+                if (value == null || !value.contains(PlotBlock.EVERYTHING) && !value
+                    .contains(block.getPlotBlock())) {
+                    return Permissions
+                        .hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false) || !(
+                        !notifyPerms || MainUtil
                             .sendMessage(player, C.FLAG_TUTORIAL_USAGE, C.FLAG_USE.s()));
                 }
                 return true;
             }
             case PLACE_BLOCK: {
-                if (plot == null) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_BUILD_ROAD.s(), notifyPerms);
-                }
                 if (!plot.hasOwner()) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_BUILD_UNOWNED.s(), notifyPerms);
+                    return Permissions
+                        .hasPermission(player, C.PERMISSION_ADMIN_BUILD_UNOWNED.s(), notifyPerms);
                 }
                 Optional<HashSet<PlotBlock>> flagValue = plot.getFlag(Flags.PLACE);
                 HashSet<PlotBlock> value;
@@ -225,21 +207,18 @@ public abstract class EventUtil {
                 } else {
                     value = null;
                 }
-                if (value == null || !value.contains(PlotBlock.EVERYTHING) && !value.contains(block.getPlotBlock())) {
-                    if (Permissions.hasPermission(player, C.PERMISSION_ADMIN_BUILD_OTHER.s(), false)) {
+                if (value == null || !value.contains(PlotBlock.EVERYTHING) && !value
+                    .contains(block.getPlotBlock())) {
+                    if (Permissions
+                        .hasPermission(player, C.PERMISSION_ADMIN_BUILD_OTHER.s(), false)) {
                         return true;
                     }
-                    return !(!notifyPerms || MainUtil.sendMessage(player, C.FLAG_TUTORIAL_USAGE, C.FLAG_PLACE.s()));
+                    return !(!notifyPerms || MainUtil
+                        .sendMessage(player, C.FLAG_TUTORIAL_USAGE, C.FLAG_PLACE.s()));
                 }
                 return true;
             }
             case TRIGGER_PHYSICAL: {
-                if (plot == null) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_ROAD.s(), false);
-                }
-                if (!plot.hasOwner()) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_UNOWNED.s(), false);
-                }
                 if (plot.getFlag(Flags.DEVICE_INTERACT).or(false)) {
                     return true;
                 }
@@ -250,8 +229,10 @@ public abstract class EventUtil {
                 } else {
                     value = null;
                 }
-                if (value == null || !value.contains(PlotBlock.EVERYTHING) && !value.contains(block.getPlotBlock())) {
-                    if (Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false)) {
+                if (value == null || !value.contains(PlotBlock.EVERYTHING) && !value
+                    .contains(block.getPlotBlock())) {
+                    if (Permissions
+                        .hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false)) {
                         return true;
                     }
                     return false;
@@ -259,12 +240,6 @@ public abstract class EventUtil {
                 return true;
             }
             case INTERACT_HANGING: {
-                if (plot == null) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_ROAD.s(), notifyPerms);
-                }
-                if (!plot.hasOwner()) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_UNOWNED.s(), notifyPerms);
-                }
                 if (plot.getFlag(Flags.HOSTILE_INTERACT).or(false)) {
                     return true;
                 }
@@ -275,21 +250,18 @@ public abstract class EventUtil {
                 } else {
                     value = null;
                 }
-                if (value == null || !value.contains(PlotBlock.EVERYTHING) && !value.contains(block.getPlotBlock())) {
-                    if (Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false)) {
+                if (value == null || !value.contains(PlotBlock.EVERYTHING) && !value
+                    .contains(block.getPlotBlock())) {
+                    if (Permissions
+                        .hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false)) {
                         return true;
                     }
-                    return !(!notifyPerms || MainUtil.sendMessage(player, C.FLAG_TUTORIAL_USAGE, C.FLAG_USE.s() + '/' + C.FLAG_HANGING_INTERACT.s()));
+                    return !(!notifyPerms || MainUtil.sendMessage(player, C.FLAG_TUTORIAL_USAGE,
+                        C.FLAG_USE.s() + '/' + C.FLAG_HANGING_INTERACT.s()));
                 }
                 return true;
             }
             case INTERACT_MISC: {
-                if (plot == null) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_ROAD.s(), notifyPerms);
-                }
-                if (!plot.hasOwner()) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_UNOWNED.s(), notifyPerms);
-                }
                 if (plot.getFlag(Flags.MISC_INTERACT).or(false)) {
                     return true;
                 }
@@ -300,21 +272,18 @@ public abstract class EventUtil {
                 } else {
                     value = null;
                 }
-                if (value == null || !value.contains(PlotBlock.EVERYTHING) && !value.contains(block.getPlotBlock())) {
-                    if (Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false)) {
+                if (value == null || !value.contains(PlotBlock.EVERYTHING) && !value
+                    .contains(block.getPlotBlock())) {
+                    if (Permissions
+                        .hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false)) {
                         return true;
                     }
-                    return !(!notifyPerms || MainUtil.sendMessage(player, C.FLAG_TUTORIAL_USAGE, C.FLAG_USE.s() + '/' + C.FLAG_MISC_INTERACT.s()));
+                    return !(!notifyPerms || MainUtil.sendMessage(player, C.FLAG_TUTORIAL_USAGE,
+                        C.FLAG_USE.s() + '/' + C.FLAG_MISC_INTERACT.s()));
                 }
                 return true;
             }
             case INTERACT_VEHICLE: {
-                if (plot == null) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_ROAD.s(), notifyPerms);
-                }
-                if (!plot.hasOwner()) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_UNOWNED.s(), notifyPerms);
-                }
                 if (plot.getFlag(Flags.VEHICLE_USE).or(false)) {
                     return true;
                 }
@@ -325,21 +294,18 @@ public abstract class EventUtil {
                 } else {
                     value = null;
                 }
-                if (value == null || !value.contains(PlotBlock.EVERYTHING) && !value.contains(block.getPlotBlock())) {
-                    if (Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false)) {
+                if (value == null || !value.contains(PlotBlock.EVERYTHING) && !value
+                    .contains(block.getPlotBlock())) {
+                    if (Permissions
+                        .hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false)) {
                         return true;
                     }
-                    return !(!notifyPerms || MainUtil.sendMessage(player, C.FLAG_TUTORIAL_USAGE, C.FLAG_USE.s() + '/' + C.FLAG_VEHICLE_USE.s()));
+                    return !(!notifyPerms || MainUtil.sendMessage(player, C.FLAG_TUTORIAL_USAGE,
+                        C.FLAG_USE.s() + '/' + C.FLAG_VEHICLE_USE.s()));
                 }
                 return true;
             }
             case SPAWN_MOB: {
-                if (plot == null) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_ROAD.s(), notifyPerms);
-                }
-                if (!plot.hasOwner()) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_UNOWNED.s(), notifyPerms);
-                }
                 if (plot.getFlag(Flags.MOB_PLACE).or(false)) {
                     return true;
                 }
@@ -350,23 +316,18 @@ public abstract class EventUtil {
                 } else {
                     value = null;
                 }
-                if (value == null || !value.contains(PlotBlock.EVERYTHING) && !value.contains(block.getPlotBlock())) {
-                    if (Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false)) {
+                if (value == null || !value.contains(PlotBlock.EVERYTHING) && !value
+                    .contains(block.getPlotBlock())) {
+                    if (Permissions
+                        .hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false)) {
                         return true;
                     }
-                    return !(!notifyPerms || MainUtil.sendMessage(player, C.FLAG_TUTORIAL_USAGE, C.FLAG_MOB_PLACE.s() + '/' + C.FLAG_PLACE.s()));
+                    return !(!notifyPerms || MainUtil.sendMessage(player, C.FLAG_TUTORIAL_USAGE,
+                        C.FLAG_MOB_PLACE.s() + '/' + C.FLAG_PLACE.s()));
                 }
                 return true;
             }
-            case PLACE_HANGING: // Handled elsewhere
-                return true;
             case PLACE_MISC: {
-                if (plot == null) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_ROAD.s(), notifyPerms);
-                }
-                if (!plot.hasOwner()) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_UNOWNED.s(), notifyPerms);
-                }
                 if (plot.getFlag(Flags.MISC_PLACE).or(false)) {
                     return true;
                 }
@@ -377,22 +338,19 @@ public abstract class EventUtil {
                 } else {
                     value = null;
                 }
-                if (value == null || !value.contains(PlotBlock.EVERYTHING) && !value.contains(block.getPlotBlock())) {
-                    if (Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false)) {
+                if (value == null || !value.contains(PlotBlock.EVERYTHING) && !value
+                    .contains(block.getPlotBlock())) {
+                    if (Permissions
+                        .hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false)) {
                         return true;
                     }
-                    return !(!notifyPerms || MainUtil.sendMessage(player, C.FLAG_TUTORIAL_USAGE, C.FLAG_MISC_PLACE.s() + '/' + C.FLAG_PLACE.s()));
+                    return !(!notifyPerms || MainUtil.sendMessage(player, C.FLAG_TUTORIAL_USAGE,
+                        C.FLAG_MISC_PLACE.s() + '/' + C.FLAG_PLACE.s()));
                 }
 
                 return true;
             }
             case PLACE_VEHICLE:
-                if (plot == null) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_ROAD.s(), notifyPerms);
-                }
-                if (!plot.hasOwner()) {
-                    return Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_UNOWNED.s(), notifyPerms);
-                }
                 if (plot.getFlag(Flags.VEHICLE_PLACE).or(false)) {
                     return true;
                 }
@@ -403,11 +361,14 @@ public abstract class EventUtil {
                 } else {
                     value = null;
                 }
-                if (value == null || !value.contains(PlotBlock.EVERYTHING) && !value.contains(block.getPlotBlock())) {
-                    if (Permissions.hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false)) {
+                if (value == null || !value.contains(PlotBlock.EVERYTHING) && !value
+                    .contains(block.getPlotBlock())) {
+                    if (Permissions
+                        .hasPermission(player, C.PERMISSION_ADMIN_INTERACT_OTHER.s(), false)) {
                         return true;
                     }
-                    return !(!notifyPerms || MainUtil.sendMessage(player, C.FLAG_TUTORIAL_USAGE, C.FLAG_VEHICLE_PLACE.s() + '/' + C.FLAG_PLACE.s()));
+                    return !(!notifyPerms || MainUtil.sendMessage(player, C.FLAG_TUTORIAL_USAGE,
+                        C.FLAG_VEHICLE_PLACE.s() + '/' + C.FLAG_PLACE.s()));
                 }
                 return true;
             default:
