@@ -35,15 +35,12 @@ public class PlayerEvents183 implements Listener {
             return;
         }
         Plot plot = area.getOwnedPlot(location);
-        if (plot == null || !plot.getFlag(Flags.EXPLOSION).or(false)) {
+        if (plot == null) {
+            event.setCancelled(true);
+        } else if (!plot.getFlag(Flags.EXPLOSION).or(false)) {
             event.setCancelled(true);
         }
-        Iterator<Block> iterator = event.blockList().iterator();
-        while (iterator.hasNext()) {
-            Block b = iterator.next();
-            if (!plot.equals(area.getOwnedPlot(BukkitUtil.getLocation(b.getLocation())))) {
-                iterator.remove();
-            }
-        }
+        event.blockList().removeIf(
+            b -> !plot.equals(area.getOwnedPlot(BukkitUtil.getLocation(b.getLocation()))));
     }
 }
