@@ -6,7 +6,6 @@ import com.intellectualcrafters.plot.object.Location;
 import com.intellectualcrafters.plot.object.PlotPlayer;
 import com.intellectualcrafters.plot.util.*;
 import com.plotsquared.bukkit.util.BukkitUtil;
-import org.apache.commons.lang.StringUtils;
 import org.bukkit.Effect;
 import org.bukkit.GameMode;
 import org.bukkit.WeatherType;
@@ -14,12 +13,11 @@ import org.bukkit.entity.Player;
 import org.bukkit.event.EventException;
 import org.bukkit.event.player.PlayerTeleportEvent;
 import org.bukkit.event.player.PlayerTeleportEvent.TeleportCause;
+import org.bukkit.permissions.PermissionAttachmentInfo;
+import org.bukkit.plugin.RegisteredListener;
 
 import java.util.Set;
 import java.util.UUID;
-
-import org.bukkit.permissions.PermissionAttachmentInfo;
-import org.bukkit.plugin.RegisteredListener;
 
 public class BukkitPlayer extends PlotPlayer {
     
@@ -47,7 +45,11 @@ public class BukkitPlayer extends PlotPlayer {
     @Override
     public Location getLocation() {
         Location location = super.getLocation();
-        return location == null ? BukkitUtil.getLocation(this.player) : location;
+        if (location == null) {
+            return BukkitUtil.getLocation(this.player);
+        } else {
+            return location;
+        }
     }
     
     @Override
@@ -153,7 +155,7 @@ public class BukkitPlayer extends PlotPlayer {
 
     @Override
     public void sendMessage(String message) {
-        if (!StringMan.isEqual(this.<String>getMeta("lastMessage"), message) || (System.currentTimeMillis() - this.<Long>getMeta("lastMessageTime") > 5000)) {
+        if (!StringMan.isEqual(this.getMeta("lastMessage"), message) || (System.currentTimeMillis() - this.<Long>getMeta("lastMessageTime") > 5000)) {
             setMeta("lastMessage", message);
             setMeta("lastMessageTime", System.currentTimeMillis());
             this.player.sendMessage(message);
