@@ -1,13 +1,13 @@
 package com.plotsquared.bukkit.titles;
 
+import org.bukkit.Bukkit;
+import org.bukkit.ChatColor;
+import org.bukkit.entity.Player;
+
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.util.HashMap;
 import java.util.Map;
-
-import org.bukkit.Bukkit;
-import org.bukkit.ChatColor;
-import org.bukkit.entity.Player;
 
 /**
  * Minecraft 1.8 Title
@@ -421,7 +421,8 @@ public class TitleManager_1_11 {
         try {
             return methodPlayerGetHandle.invoke(player);
         } catch (Exception e) {
-            e.printStackTrace();
+            //don't print a stacktrace. It just encourages uninformed users to report non-bugs.
+            //e.printStackTrace();
             return null;
         }
     }
