@@ -6,7 +6,9 @@ import com.intellectualcrafters.plot.object.PlotBlock;
 import com.intellectualcrafters.plot.util.MainUtil;
 import com.intellectualcrafters.plot.util.block.ScopedLocalBlockQueue;
 import com.plotsquared.bukkit.util.BukkitUtil;
+
 import java.util.Arrays;
+
 import org.bukkit.Chunk;
 import org.bukkit.Material;
 import org.bukkit.World;
@@ -17,7 +19,7 @@ import org.bukkit.material.MaterialData;
 
 public class GenChunk extends ScopedLocalBlockQueue {
 
-    public final Biome[] biomes;
+    private final Biome[] biomes;
     public short[][] result;
     public byte[][] result_data;
     public ChunkData cd;
@@ -25,8 +27,8 @@ public class GenChunk extends ScopedLocalBlockQueue {
 
     public Chunk chunk;
     public String world;
-    public int cx;
-    public int cz;
+    private int cx;
+    private int cz;
 
     public GenChunk(Chunk chunk, ChunkWrapper wrap) {
         super(null, new Location(null, 0, 0, 0), new Location(null, 15, 255, 15));
@@ -61,8 +63,7 @@ public class GenChunk extends ScopedLocalBlockQueue {
         return new ChunkWrapper(chunk.getWorld().getName(), chunk.getX(), chunk.getZ());
     }
 
-    @Override
-    public void fillBiome(String biomeName) {
+    @Override public void fillBiome(String biomeName) {
         if (grid == null) {
             return;
         }
@@ -74,9 +75,9 @@ public class GenChunk extends ScopedLocalBlockQueue {
         }
     }
 
-    @Override
-    public void setCuboid(Location pos1, Location pos2, PlotBlock block) {
-        if (block.data == 0 && result != null && pos1.getX() == 0 && pos1.getZ() == 0 && pos2.getX() == 15 && pos2.getZ() == 15) {
+    @Override public void setCuboid(Location pos1, Location pos2, PlotBlock block) {
+        if (block.data == 0 && result != null && pos1.getX() == 0 && pos1.getZ() == 0
+            && pos2.getX() == 15 && pos2.getZ() == 15) {
             for (int y = pos1.getY(); y <= pos2.getY(); y++) {
                 int layer = y >> 4;
                 short[] data = result[layer];
@@ -92,8 +93,7 @@ public class GenChunk extends ScopedLocalBlockQueue {
         }
     }
 
-    @Override
-    public boolean setBiome(int x, int z, String biome) {
+    @Override public boolean setBiome(int x, int z, String biome) {
         return setBiome(x, z, Biome.valueOf(biome.toUpperCase()));
     }
 
@@ -113,10 +113,13 @@ public class GenChunk extends ScopedLocalBlockQueue {
         return false;
     }
 
-    @Override
-    public boolean setBlock(int x, int y, int z, int id, int data) {
+    @Override public boolean setBlock(int x, int y, int z, int id, int data) {
         if (this.result == null) {
-            this.cd.setBlock(x, y, z, new MaterialData(Material.getMaterial(id), (byte) data));
+            Material material = Material.getMaterial(id);
+            if (material == null) {
+                return false;
+            }
+            this.cd.setBlock(x, y, z, new MaterialData(material, (byte) data));
             return true;
         }
         int i = MainUtil.CACHE_I[y][x][z];
@@ -136,8 +139,7 @@ public class GenChunk extends ScopedLocalBlockQueue {
         return true;
     }
 
-    @Override
-    public PlotBlock getBlock(int x, int y, int z) {
+    @Override public PlotBlock getBlock(int x, int y, int z) {
         int i = MainUtil.CACHE_I[y][x][z];
         if (result == null) {
             MaterialData md = cd.getTypeAndData(x, y, z);
@@ -160,30 +162,40 @@ public class GenChunk extends ScopedLocalBlockQueue {
     }
 
     public int getX() {
-        return chunk == null ? cx : chunk.getX();
+        if (chunk == null) {
+            return cx;
+        } else {
+            return chunk.getX();
+        }
     }
 
     public int getZ() {
-        return chunk == null ? cz : chunk.getZ();
+        if (chunk == null) {
+            return cz;
+        } else {
+            return chunk.getZ();
+        }
     }
 
-    @Override
-    public String getWorld() {
-        return chunk == null ? world : chunk.getWorld().getName();
+    @Override public String getWorld() {
+        if (chunk == null) {
+            return world;
+        } else {
+            return chunk.getWorld().getName();
+        }
     }
 
-    @Override
-    public Location getMax() {
+    @Override public Location getMax() {
         return new Location(getWorld(), 15 + (getX() << 4), 255, 15 + (getZ() << 4));
     }
 
-    @Override
-    public Location getMin() {
+    @Override public Location getMin() {
         return new Location(getWorld(), getX() << 4, 0, getZ() << 4);
     }
 
     public GenChunk clone() {
-        GenChunk toReturn = new GenChunk(chunk, new ChunkWrapper(getWorld(), chunk.getX(), chunk.getZ()));
+        GenChunk toReturn =
+            new GenChunk(chunk, new ChunkWrapper(getWorld(), chunk.getX(), chunk.getZ()));
         if (this.result != null) {
             for (int i = 0; i < this.result.length; i++) {
                 short[] matrix = this.result[i];
