@@ -25,9 +25,9 @@ public class HybridGen extends IndependentPlotGenerator {
     }
 
     private void placeSchem(HybridPlotWorld world, ScopedLocalBlockQueue result, short relativeX, short relativeZ, int x, int z, boolean isRoad) {
-        int minY;
-        if(isRoad || Settings.Schematics.PASTE_ON_TOP) {
-            minY = Math.min(world.PLOT_HEIGHT, world.ROAD_HEIGHT);
+        int minY; // Math.min(world.PLOT_HEIGHT, world.ROAD_HEIGHT);
+        if (isRoad || Settings.Schematics.PASTE_ON_TOP) {
+            minY = world.SCHEM_Y;
         } else {
             minY = 1;
         }
