@@ -1,48 +1,50 @@
 package com.intellectualcrafters.plot.generator;
 
+import com.google.common.collect.Sets;
 import com.intellectualcrafters.plot.PS;
 import com.intellectualcrafters.plot.commands.Template;
 import com.intellectualcrafters.plot.config.Settings;
-import com.intellectualcrafters.plot.object.FileBytes;
-import com.intellectualcrafters.plot.object.Location;
-import com.intellectualcrafters.plot.object.Plot;
-import com.intellectualcrafters.plot.object.PlotArea;
-import com.intellectualcrafters.plot.object.PlotBlock;
-import com.intellectualcrafters.plot.object.PlotId;
-import com.intellectualcrafters.plot.object.RunnableVal;
+import com.intellectualcrafters.plot.object.*;
 import com.intellectualcrafters.plot.util.ChunkManager;
 import com.intellectualcrafters.plot.util.MainUtil;
 import com.intellectualcrafters.plot.util.MathMan;
 import com.intellectualcrafters.plot.util.block.GlobalBlockQueue;
 import com.intellectualcrafters.plot.util.block.LocalBlockQueue;
+
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
-import java.util.Collections;
 import java.util.HashSet;
 
 public class HybridPlotManager extends ClassicPlotManager {
 
     public static boolean REGENERATIVE_CLEAR = true;
 
-    @Override
-    public void exportTemplate(PlotArea plotArea) throws IOException {
-        HashSet<FileBytes> files = new HashSet<>(
-                Collections.singletonList(new FileBytes(Settings.Paths.TEMPLATES + "/tmp-data.yml", Template.getBytes(plotArea))));
-        String dir = "schematics" + File.separator + "GEN_ROAD_SCHEMATIC" + File.separator + plotArea.worldname + File.separator;
-        String newDir = "schematics" + File.separator + "GEN_ROAD_SCHEMATIC" + File.separator + "__TEMP_DIR__" + File.separator;
+    @Override public void exportTemplate(PlotArea plotArea) throws IOException {
+        HashSet<FileBytes> files = Sets.newHashSet(
+            new FileBytes(Settings.Paths.TEMPLATES + "/tmp-data.yml", Template.getBytes(plotArea)));
+        String dir = "schematics" + File.separator + "GEN_ROAD_SCHEMATIC" + File.separator
+            + plotArea.worldname + File.separator;
+        String newDir =
+            "schematics" + File.separator + "GEN_ROAD_SCHEMATIC" + File.separator + "__TEMP_DIR__"
+                + File.separator;
         try {
-            File sideroad = MainUtil.getFile(PS.get().IMP.getDirectory(), dir + "sideroad.schematic");
+            File sideroad =
+                MainUtil.getFile(PS.get().IMP.getDirectory(), dir + "sideroad.schematic");
             if (sideroad.exists()) {
-                files.add(new FileBytes(newDir + "sideroad.schematic", Files.readAllBytes(sideroad.toPath())));
+                files.add(new FileBytes(newDir + "sideroad.schematic",
+                    Files.readAllBytes(sideroad.toPath())));
             }
-            File intersection = MainUtil.getFile(PS.get().IMP.getDirectory(), "intersection.schematic");
+            File intersection =
+                MainUtil.getFile(PS.get().IMP.getDirectory(), "intersection.schematic");
             if (intersection.exists()) {
-                files.add(new FileBytes(newDir + "intersection.schematic", Files.readAllBytes(intersection.toPath())));
+                files.add(new FileBytes(newDir + "intersection.schematic",
+                    Files.readAllBytes(intersection.toPath())));
             }
             File plot = MainUtil.getFile(PS.get().IMP.getDirectory(), dir + "plot.schematic");
             if (plot.exists()) {
-                files.add(new FileBytes(newDir + "plot.schematic", Files.readAllBytes(plot.toPath())));
+                files.add(
+                    new FileBytes(newDir + "plot.schematic", Files.readAllBytes(plot.toPath())));
             }
         } catch (IOException e) {
             e.printStackTrace();
@@ -50,8 +52,7 @@ public class HybridPlotManager extends ClassicPlotManager {
         Template.zipAll(plotArea.worldname, files);
     }
 
-    @Override
-    public boolean createRoadEast(PlotArea plotArea, Plot plot) {
+    @Override public boolean createRoadEast(PlotArea plotArea, Plot plot) {
         super.createRoadEast(plotArea, plot);
         HybridPlotWorld hpw = (HybridPlotWorld) plotArea;
         PlotId id = plot.getId();
@@ -59,7 +60,9 @@ public class HybridPlotManager extends ClassicPlotManager {
         Location bot = getPlotBottomLocAbs(hpw, id2);
         Location top = getPlotTopLocAbs(hpw, id);
         Location pos1 = new Location(plotArea.worldname, top.getX() + 1, 0, bot.getZ() - 1);
-        Location pos2 = new Location(plotArea.worldname, bot.getX(), Math.min(getWorldHeight(), 255), top.getZ() + 1);
+        Location pos2 =
+            new Location(plotArea.worldname, bot.getX(), Math.min(getWorldHeight(), 255),
+                top.getZ() + 1);
         MainUtil.resetBiome(plotArea, pos1, pos2);
         if (!hpw.ROAD_SCHEMATIC_ENABLED) {
             return true;
@@ -70,9 +73,62 @@ public class HybridPlotManager extends ClassicPlotManager {
         return true;
     }
 
-    private void createSchemAbs(HybridPlotWorld hpw, LocalBlockQueue queue, Location pos1, Location pos2, boolean clear) {
+    @Override public boolean createRoadSouth(PlotArea plotArea, Plot plot) {
+        super.createRoadSouth(plotArea, plot);
+        HybridPlotWorld hpw = (HybridPlotWorld) plotArea;
+        PlotId id = plot.getId();
+        PlotId id2 = new PlotId(id.x, id.y + 1);
+        Location bot = getPlotBottomLocAbs(hpw, id2);
+        Location top = getPlotTopLocAbs(hpw, id);
+        Location pos1 = new Location(plotArea.worldname, bot.getX() - 1, 0, top.getZ() + 1);
+        Location pos2 =
+            new Location(plotArea.worldname, top.getX() + 1, Math.min(getWorldHeight(), 255),
+                bot.getZ());
+        MainUtil.resetBiome(plotArea, pos1, pos2);
+        if (!hpw.ROAD_SCHEMATIC_ENABLED) {
+            return true;
+        }
+        LocalBlockQueue queue = hpw.getQueue(false);
+        createSchemAbs(hpw, queue, pos1, pos2, true);
+        queue.enqueue();
+        return true;
+    }
+
+    @Override public boolean createRoadSouthEast(PlotArea plotArea, Plot plot) {
+        super.createRoadSouthEast(plotArea, plot);
+        HybridPlotWorld hpw = (HybridPlotWorld) plotArea;
+        PlotId id = plot.getId();
+        PlotId id2 = new PlotId(id.x + 1, id.y + 1);
+        Location pos1 = getPlotTopLocAbs(hpw, id).add(1, 0, 1);
+        Location pos2 = getPlotBottomLocAbs(hpw, id2);
+        pos1.setY(0);
+        pos2.setY(Math.min(getWorldHeight(), 255));
+        LocalBlockQueue queue = hpw.getQueue(false);
+        createSchemAbs(hpw, queue, pos1, pos2, true);
+        if (hpw.ROAD_SCHEMATIC_ENABLED) {
+            createSchemAbs(hpw, queue, pos1, pos2, true);
+        }
+        queue.enqueue();
+        return true;
+    }
+
+    /**
+     * Remove sign for a plot.
+     */
+    @Override public Location getSignLoc(PlotArea plotArea, Plot plot) {
+        HybridPlotWorld dpw = (HybridPlotWorld) plotArea;
+        return dpw.getSignLocation(plot);
+    }
+
+    private void createSchemAbs(HybridPlotWorld hpw, LocalBlockQueue queue, Location pos1,
+        Location pos2, boolean clear) {
         int size = hpw.SIZE;
-        int minY = Math.min(hpw.PLOT_HEIGHT, hpw.ROAD_HEIGHT);
+        int minY;
+        if (Settings.Schematics.PASTE_ON_TOP) {
+            minY = hpw.SCHEM_Y;
+        } else {
+            minY = 1;
+        }
         for (int x = pos1.getX(); x <= pos2.getX(); x++) {
             short absX = (short) ((x - hpw.ROAD_OFFSET_X) % size);
             if (absX < 0) {
@@ -96,59 +152,21 @@ public class HybridPlotManager extends ClassicPlotManager {
         }
     }
 
-    @Override
-    public boolean createRoadSouth(PlotArea plotArea, Plot plot) {
-        super.createRoadSouth(plotArea, plot);
-        HybridPlotWorld hpw = (HybridPlotWorld) plotArea;
-        PlotId id = plot.getId();
-        PlotId id2 = new PlotId(id.x, id.y + 1);
-        Location bot = getPlotBottomLocAbs(hpw, id2);
-        Location top = getPlotTopLocAbs(hpw, id);
-        Location pos1 = new Location(plotArea.worldname, bot.getX() - 1, 0, top.getZ() + 1);
-        Location pos2 = new Location(plotArea.worldname, top.getX() + 1, Math.min(getWorldHeight(), 255), bot.getZ());
-        MainUtil.resetBiome(plotArea, pos1, pos2);
-        if (!hpw.ROAD_SCHEMATIC_ENABLED) {
-            return true;
-        }
-        LocalBlockQueue queue = hpw.getQueue(false);
-        createSchemAbs(hpw, queue, pos1, pos2, true);
-        queue.enqueue();
-        return true;
-    }
-
-    @Override
-    public boolean createRoadSouthEast(PlotArea plotArea, Plot plot) {
-        super.createRoadSouthEast(plotArea, plot);
-        HybridPlotWorld hpw = (HybridPlotWorld) plotArea;
-        PlotId id = plot.getId();
-        PlotId id2 = new PlotId(id.x + 1, id.y + 1);
-        Location pos1 = getPlotTopLocAbs(hpw, id).add(1, 0, 1);
-        Location pos2 = getPlotBottomLocAbs(hpw, id2);
-        pos1.setY(0);
-        pos2.setY(Math.min(getWorldHeight(), 255));
-        LocalBlockQueue queue = hpw.getQueue(false);
-        createSchemAbs(hpw, queue, pos1, pos2, true);
-        if (hpw.ROAD_SCHEMATIC_ENABLED) {
-            createSchemAbs(hpw, queue, pos1, pos2, true);
-        }
-        queue.enqueue();
-        return true;
-    }
-
     /**
      * <p>Clearing the plot needs to only consider removing the blocks - This implementation has used the setCuboidAsync
      * function, as it is fast, and uses NMS code - It also makes use of the fact that deleting chunks is a lot faster
      * than block updates This code is very messy, but you don't need to do something quite as complex unless you happen
      * to have 512x512 sized plots. </p>
      */
-    @Override
-    public boolean clearPlot(final PlotArea plotArea, Plot plot, final Runnable whenDone) {
+    @Override public boolean clearPlot(final PlotArea plotArea, Plot plot,
+        final Runnable whenDone) {
         final String world = plotArea.worldname;
         final HybridPlotWorld dpw = (HybridPlotWorld) plotArea;
         Location pos1 = plot.getBottomAbs();
         Location pos2 = plot.getExtendedTopAbs();
         // If augmented
-        final boolean canRegen = (plotArea.TYPE == 0) && (plotArea.TERRAIN == 0) && REGENERATIVE_CLEAR;
+        final boolean canRegen =
+            (plotArea.TYPE == 0) && (plotArea.TERRAIN == 0) && REGENERATIVE_CLEAR;
         // The component blocks
         final PlotBlock[] plotfloor = dpw.TOP_BLOCK;
         final PlotBlock[] filling = dpw.MAIN_BLOCK;
@@ -162,8 +180,7 @@ public class HybridPlotManager extends ClassicPlotManager {
         final String biome = dpw.PLOT_BIOME;
         final LocalBlockQueue queue = plotArea.getQueue(false);
         ChunkManager.chunkTask(pos1, pos2, new RunnableVal<int[]>() {
-            @Override
-            public void run(int[] value) {
+            @Override public void run(int[] value) {
                 // If the chunk isn't near the edge and it isn't an augmented world we can just regen the whole chunk
                 if (canRegen && (value[6] == 0)) {
                     queue.regenChunk(value[0], value[1]);
@@ -192,8 +209,7 @@ public class HybridPlotManager extends ClassicPlotManager {
                 pastePlotSchematic(dpw, queue, bot, top);
             }
         }, new Runnable() {
-            @Override
-            public void run() {
+            @Override public void run() {
                 queue.enqueue();
                 // And notify whatever called this when plot clearing is done
                 GlobalBlockQueue.IMP.addTask(whenDone);
@@ -202,19 +218,11 @@ public class HybridPlotManager extends ClassicPlotManager {
         return true;
     }
 
-    public void pastePlotSchematic(HybridPlotWorld plotWorld, LocalBlockQueue queue, Location l1, Location l2) {
+    public void pastePlotSchematic(HybridPlotWorld plotWorld, LocalBlockQueue queue, Location l1,
+        Location l2) {
         if (!plotWorld.PLOT_SCHEMATIC) {
             return;
         }
         createSchemAbs(plotWorld, queue, l1, l2, false);
-    }
-
-    /**
-     * Remove sign for a plot.
-     */
-    @Override
-    public Location getSignLoc(PlotArea plotArea, Plot plot) {
-        HybridPlotWorld dpw = (HybridPlotWorld) plotArea;
-        return dpw.getSignLocation(plot);
     }
 }
