@@ -1,19 +1,20 @@
 package com.intellectualcrafters.plot.object.worlds;
 
+import com.google.common.collect.Sets;
 import com.intellectualcrafters.plot.config.C;
 import com.intellectualcrafters.plot.flag.Flag;
-import com.intellectualcrafters.plot.object.BlockLoc;
-import com.intellectualcrafters.plot.object.Location;
-import com.intellectualcrafters.plot.object.Plot;
-import com.intellectualcrafters.plot.object.PlotArea;
-import com.intellectualcrafters.plot.object.PlotId;
-import com.intellectualcrafters.plot.object.PlotPlayer;
-import com.intellectualcrafters.plot.object.RegionWrapper;
+import com.intellectualcrafters.plot.object.*;
+
+import javax.annotation.Nonnull;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.UUID;
 
 public class SinglePlot extends Plot {
+    private HashSet<RegionWrapper> regions = Sets.newHashSet(
+        new RegionWrapper(Integer.MIN_VALUE, Integer.MAX_VALUE, Integer.MIN_VALUE,
+            Integer.MAX_VALUE));
+
     public SinglePlot(PlotArea area, PlotId id, UUID owner) {
         super(area, id, owner);
     }
@@ -26,18 +27,32 @@ public class SinglePlot extends Plot {
         super(area, id, owner, temp);
     }
 
-    public SinglePlot(PlotId id, UUID owner, HashSet<UUID> trusted, HashSet<UUID> members, HashSet<UUID> denied, String alias, BlockLoc position, Collection<Flag> flags, PlotArea area, boolean[] merged, long timestamp, int temp) {
-        super(id, owner, trusted, members, denied, alias, position, flags, area, merged, timestamp, temp);
+    public SinglePlot(PlotId id, UUID owner, HashSet<UUID> trusted, HashSet<UUID> members,
+        HashSet<UUID> denied, String alias, BlockLoc position, Collection<Flag> flags,
+        PlotArea area, boolean[] merged, long timestamp, int temp) {
+        super(id, owner, trusted, members, denied, alias, position, flags, area, merged, timestamp,
+            temp);
     }
 
-    @Override
-    public String getWorldName() {
+    @Override public String getWorldName() {
         return getId().toCommaSeparatedString();
     }
 
-    @Override
-    public SinglePlotArea getArea() {
+    @Override public SinglePlotArea getArea() {
         return (SinglePlotArea) super.getArea();
+    }
+
+    @Override protected boolean isLoaded() {
+        getArea().loadWorld(getId());
+        return super.isLoaded();
+    }
+
+    @Override public Location getSide() {
+        return getCenter();
+    }
+
+    @Nonnull @Override public HashSet<RegionWrapper> getRegions() {
+        return regions;
     }
 
     public boolean teleportPlayer(final PlotPlayer player) {
@@ -47,27 +62,6 @@ public class SinglePlot extends Plot {
             C.NOT_LOADED.send(player);
             return false;
         }
-    }
-
-    @Override
-    public Location getSide() {
-        return getCenter();
-    }
-
-    @Override
-    protected boolean isLoaded() {
-        getArea().loadWorld(getId());
-        return super.isLoaded();
-    }
-    private HashSet<RegionWrapper> regions;
-    {
-        regions = new HashSet<>();
-        regions.add(new RegionWrapper(Integer.MIN_VALUE, Integer.MAX_VALUE, Integer.MIN_VALUE, Integer.MAX_VALUE));
-    }
-
-    @Override
-    public HashSet<RegionWrapper> getRegions() {
-        return regions;
     }
 
     // getCenter getSide getHome getDefaultHome getBiome
