@@ -156,22 +156,19 @@ public class PlotListener {
             }
             if (titles) {
                 if (!C.TITLE_ENTERED_PLOT.s().isEmpty() || !C.TITLE_ENTERED_PLOT_SUB.s().isEmpty()) {
-                    TaskManager.runTaskLaterAsync(new Runnable() {
-                        @Override
-                        public void run() {
-                            Plot lastPlot = player.getMeta("lastplot");
-                            if ((lastPlot != null) && plot.getId().equals(lastPlot.getId())) {
-                                Map<String, String> replacements = new HashMap<>();
-                                replacements.put("%x%", String.valueOf(lastPlot.getId().x));
-                                replacements.put("%z%", lastPlot.getId().y + "");
-                                replacements.put("%world%", plot.getArea().toString());
-                                replacements.put("%greeting%", greeting);
-                                replacements.put("%alias", plot.toString());
-                                replacements.put("%s", MainUtil.getName(plot.owner));
-                                String main = StringMan.replaceFromMap(C.TITLE_ENTERED_PLOT.s(), replacements);
-                                String sub = StringMan.replaceFromMap(C.TITLE_ENTERED_PLOT_SUB.s(), replacements);
-                                AbstractTitle.sendTitle(player, main, sub);
-                            }
+                    TaskManager.runTaskLaterAsync(() -> {
+                        Plot lastPlot = player.getMeta("lastplot");
+                        if ((lastPlot != null) && plot.getId().equals(lastPlot.getId())) {
+                            Map<String, String> replacements = new HashMap<>();
+                            replacements.put("%x%", String.valueOf(lastPlot.getId().x));
+                            replacements.put("%z%", lastPlot.getId().y + "");
+                            replacements.put("%world%", plot.getArea().toString());
+                            replacements.put("%greeting%", greeting);
+                            replacements.put("%alias", plot.toString());
+                            replacements.put("%s", MainUtil.getName(plot.owner));
+                            String main = StringMan.replaceFromMap(C.TITLE_ENTERED_PLOT.s(), replacements);
+                            String sub = StringMan.replaceFromMap(C.TITLE_ENTERED_PLOT_SUB.s(), replacements);
+                            AbstractTitle.sendTitle(player, main, sub);
                         }
                     }, 20);
                 }
