@@ -31,7 +31,6 @@ import com.plotsquared.sponge.util.block.SpongeLocalQueue;
 import com.plotsquared.sponge.uuid.SpongeLowerOfflineUUIDWrapper;
 import com.plotsquared.sponge.uuid.SpongeOnlineUUIDWrapper;
 import com.plotsquared.sponge.uuid.SpongeUUIDHandler;
-import net.minecrell.mcstats.SpongeStatsLite;
 import org.slf4j.Logger;
 import org.spongepowered.api.Game;
 import org.spongepowered.api.Server;
@@ -75,9 +74,6 @@ public class SpongeMain implements IPlotMain {
     @Inject
     private Game game;
 
-    @Inject
-    public SpongeStatsLite stats;
-
     private Server server;
 
     @Inject
@@ -111,8 +107,8 @@ public class SpongeMain implements IPlotMain {
 
     @Listener
     public void onPreInitialize(GamePreInitializationEvent event) {
-        getLogger().info("The metrics section in PlotSquared is ignored in favor of the actual metrics reporter configurations.");
-        this.stats.start();
+        //getLogger().info("The metrics section in PlotSquared is ignored in favor of the actual metrics reporter configurations.");
+        //this.stats.start();
     }
 
     @Listener
@@ -123,12 +119,7 @@ public class SpongeMain implements IPlotMain {
         this.game.getRegistry().register(WorldGeneratorModifier.class, (WorldGeneratorModifier) PS.get().IMP.getDefaultGenerator().specify(null));
         this.game.getRegistry().register(WorldGeneratorModifier.class, (WorldGeneratorModifier) new SingleWorldGenerator().specify(null));
         if (Settings.Enabled_Components.WORLDS) {
-            TaskManager.IMP.taskRepeat(new Runnable() {
-                @Override
-                public void run() {
-                    unload();
-                }
-            }, 20);
+            TaskManager.IMP.taskRepeat(this::unload, 20);
         }
     }
 
@@ -216,7 +207,12 @@ public class SpongeMain implements IPlotMain {
         PS.log("Checking minecraft version: Sponge: ");
         String version = this.game.getPlatform().getMinecraftVersion().getName();
         String[] split = version.split("\\.");
-        return new int[]{Integer.parseInt(split[0]), Integer.parseInt(split[1]), split.length == 3 ? Integer.parseInt(split[2]) : 0};
+        if (split.length == 3) {
+            return new int[] {Integer.parseInt(split[0]), Integer.parseInt(split[1]),
+                Integer.parseInt(split[2])};
+        } else {
+            return new int[] {Integer.parseInt(split[0]), Integer.parseInt(split[1]), 0};
+        }
     }
 
     @Override
