@@ -13,8 +13,6 @@ import org.spongepowered.api.data.DataContainer;
 import org.spongepowered.api.world.biome.BiomeGenerationSettings;
 import org.spongepowered.api.world.biome.BiomeType;
 import org.spongepowered.api.world.biome.BiomeTypes;
-import org.spongepowered.api.world.extent.MutableBiomeVolume;
-import org.spongepowered.api.world.gen.BiomeGenerator;
 import org.spongepowered.api.world.gen.GenerationPopulator;
 import org.spongepowered.api.world.gen.WorldGenerator;
 import org.spongepowered.api.world.gen.WorldGeneratorModifier;
@@ -72,22 +70,19 @@ public class SpongePlotGenerator implements WorldGeneratorModifier, GeneratorWra
     public void modifyWorldGenerator(WorldProperties world, DataContainer settings, WorldGenerator worldGenerator) {
         String worldName = world.getWorldName();
         worldGenerator.setBaseGenerationPopulator(new SpongeTerrainGen(this.plotGenerator));
-        worldGenerator.setBiomeGenerator(new BiomeGenerator() {
-            @Override
-            public void generateBiomes(MutableBiomeVolume buffer) {
-                PlotArea area = PS.get().getPlotArea(worldName, null);
-                if (area != null) {
-                    BiomeType biome = SpongeUtil.getBiome(area.PLOT_BIOME);
-                    Vector3i min = buffer.getBiomeMin();
-                    Vector3i max = buffer.getBiomeMax();
-                    for (int x = min.getX(); x <= max.getX(); x++) {
-                        for (int z = min.getZ(); z <= max.getZ(); z++) {
-                            buffer.setBiome(x, 0, z, biome);
-                    }
+        worldGenerator.setBiomeGenerator(buffer -> {
+            PlotArea area = PS.get().getPlotArea(worldName, null);
+            if (area != null) {
+                BiomeType biome = SpongeUtil.getBiome(area.PLOT_BIOME);
+                Vector3i min = buffer.getBiomeMin();
+                Vector3i max = buffer.getBiomeMax();
+                for (int x = min.getX(); x <= max.getX(); x++) {
+                    for (int z = min.getZ(); z <= max.getZ(); z++) {
+                        buffer.setBiome(x, 0, z, biome);
                 }
-                }
-        }
-        });
+            }
+            }
+    });
         for (BiomeType type : ReflectionUtils.<BiomeType> getStaticFields(BiomeTypes.class)) {
             BiomeGenerationSettings biomeSettings = worldGenerator.getBiomeSettings(type);
             biomeSettings.getGenerationPopulators().clear();
