@@ -6,10 +6,8 @@ import com.intellectualcrafters.plot.object.PlotPlayer;
 import com.intellectualcrafters.plot.util.UUIDHandler;
 import com.intellectualcrafters.plot.uuid.UUIDWrapper;
 import com.plotsquared.sponge.SpongeMain;
-import org.spongepowered.api.entity.living.player.Player;
 import org.spongepowered.api.profile.GameProfile;
 
-import java.util.Collection;
 import java.util.UUID;
 
 public class SpongeLowerOfflineUUIDWrapper extends UUIDWrapper {
@@ -76,12 +74,7 @@ public class SpongeLowerOfflineUUIDWrapper extends UUIDWrapper {
             }
         };
     }
-    
-    public Player[] getOnlinePlayers() {
-        Collection<Player> onlinePlayers = SpongeMain.THIS.getServer().getOnlinePlayers();
-        return onlinePlayers.toArray(new Player[onlinePlayers.size()]);
-    }
-    
+
     @Override
     public UUID getUUID(String name) {
         return UUID.nameUUIDFromBytes(("OfflinePlayer:" + name.toLowerCase()).getBytes(Charsets.UTF_8));
